Remove Druid implem and change file organization
