--- SOURCES/include/bias_corrected_estimate.hpp
@@ -1,36 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
-#ifndef _BIAS_CORRECTED_ESTIMATE_H_
-#define _BIAS_CORRECTED_ESTIMATE_H_
-
-#include <vector>
-
-class BiasCorrectedEstimate {
-  static double kNeighborsInterpolationBias(uint64_t rawEstimate, uint8_t precision);
-
-  static const std::vector<std::vector<double>> rawEstimateData;
-  static const std::vector<std::vector<double>> biasData;
-
-public:
-  static uint64_t estimate(uint64_t rawEstimate, uint8_t precision);
-};
-
-
-#endif
