--- SOURCES/include/druid_precalc_lookups.hpp
@@ -1,112 +1,0 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
-static constexpr int positionOf1Lookup[256] = {0,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,6,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,7,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,6,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,8,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,6,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,7,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,6,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1,5,1,2,1,3,1,2,1,4,1,2,1,3,1,2,1};
-
-/*
-  // we have to keep track of the number of zeroes in each of the two halves of the byte register (0, 1, or 2)
-  private final static int[] numZeroLookup = new int[256];
-
-  static {
-    for (int i = 0; i < numZeroLookup.length; ++i) {
-      numZeroLookup[i] = (((i & 0xf0) == 0) ? 1 : 0) + (((i & 0x0f) == 0) ? 1 : 0);
-    }
-  }
-*/
-
-static constexpr int numZeroLookup[256] = {2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,};
-
-/**
-  private final static double[][] minNumRegisterLookup = new double[64][256];
-
-  static {
-    for (int registerOffset = 0; registerOffset < 64; ++registerOffset) {
-      for (int register = 0; register < 256; ++register) {
-        final int upper = ((register & 0xf0) >> 4) + registerOffset;
-        final int lower = (register & 0x0f) + registerOffset;
-        minNumRegisterLookup[registerOffset][register] = 1.0d / Math.pow(2, upper) + 1.0d / Math.pow(2, lower);
-      }
-    }
-  }
-*/
-
-static constexpr double minNumRegisterLookup[64][256] = {{2,1.5,1.25,1.125,1.0625,1.03125,1.01562,1.00781,1.00391,1.00195,1.00098,1.00049,1.00024,1.00012,1.00006,1.00003,1.5,1,0.75,0.625,0.5625,0.53125,0.515625,0.507812,0.503906,0.501953,0.500977,0.500488,0.500244,0.500122,0.500061,0.500031,1.25,0.75,0.5,0.375,0.3125,0.28125,0.265625,0.257812,0.253906,0.251953,0.250977,0.250488,0.250244,0.250122,0.250061,0.250031,1.125,0.625,0.375,0.25,0.1875,0.15625,0.140625,0.132812,0.128906,0.126953,0.125977,0.125488,0.125244,0.125122,0.125061,0.125031,1.0625,0.5625,0.3125,0.1875,0.125,0.09375,0.078125,0.0703125,0.0664062,0.0644531,0.0634766,0.0629883,0.0627441,0.0626221,0.062561,0.0625305,1.03125,0.53125,0.28125,0.15625,0.09375,0.0625,0.046875,0.0390625,0.0351562,0.0332031,0.0322266,0.0317383,0.0314941,0.0313721,0.031311,0.0312805,1.01562,0.515625,0.265625,0.140625,0.078125,0.046875,0.03125,0.0234375,0.0195312,0.0175781,0.0166016,0.0161133,0.0158691,0.0157471,0.015686,0.0156555,1.00781,0.507812,0.257812,0.132812,0.0703125,0.0390625,0.0234375,0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,1.00391,0.503906,0.253906,0.128906,0.0664062,0.0351562,0.0195312,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,1.00195,0.501953,0.251953,0.126953,0.0644531,0.0332031,0.0175781,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,1.00098,0.500977,0.250977,0.125977,0.0634766,0.0322266,0.0166016,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,1.00049,0.500488,0.250488,0.125488,0.0629883,0.0317383,0.0161133,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,1.00024,0.500244,0.250244,0.125244,0.0627441,0.0314941,0.0158691,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,1.00012,0.500122,0.250122,0.125122,0.0626221,0.0313721,0.0157471,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,1.00006,0.500061,0.250061,0.125061,0.062561,0.031311,0.015686,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,1.00003,0.500031,0.250031,0.125031,0.0625305,0.0312805,0.0156555,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,},
-{1,0.75,0.625,0.5625,0.53125,0.515625,0.507812,0.503906,0.501953,0.500977,0.500488,0.500244,0.500122,0.500061,0.500031,0.500015,0.75,0.5,0.375,0.3125,0.28125,0.265625,0.257812,0.253906,0.251953,0.250977,0.250488,0.250244,0.250122,0.250061,0.250031,0.250015,0.625,0.375,0.25,0.1875,0.15625,0.140625,0.132812,0.128906,0.126953,0.125977,0.125488,0.125244,0.125122,0.125061,0.125031,0.125015,0.5625,0.3125,0.1875,0.125,0.09375,0.078125,0.0703125,0.0664062,0.0644531,0.0634766,0.0629883,0.0627441,0.0626221,0.062561,0.0625305,0.0625153,0.53125,0.28125,0.15625,0.09375,0.0625,0.046875,0.0390625,0.0351562,0.0332031,0.0322266,0.0317383,0.0314941,0.0313721,0.031311,0.0312805,0.0312653,0.515625,0.265625,0.140625,0.078125,0.046875,0.03125,0.0234375,0.0195312,0.0175781,0.0166016,0.0161133,0.0158691,0.0157471,0.015686,0.0156555,0.0156403,0.507812,0.257812,0.132812,0.0703125,0.0390625,0.0234375,0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,0.00782776,0.503906,0.253906,0.128906,0.0664062,0.0351562,0.0195312,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.501953,0.251953,0.126953,0.0644531,0.0332031,0.0175781,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.500977,0.250977,0.125977,0.0634766,0.0322266,0.0166016,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.500488,0.250488,0.125488,0.0629883,0.0317383,0.0161133,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.500244,0.250244,0.125244,0.0627441,0.0314941,0.0158691,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.500122,0.250122,0.125122,0.0626221,0.0313721,0.0157471,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.500061,0.250061,0.125061,0.062561,0.031311,0.015686,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,0.500031,0.250031,0.125031,0.0625305,0.0312805,0.0156555,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,0.500015,0.250015,0.125015,0.0625153,0.0312653,0.0156403,0.00782776,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,},
-{0.5,0.375,0.3125,0.28125,0.265625,0.257812,0.253906,0.251953,0.250977,0.250488,0.250244,0.250122,0.250061,0.250031,0.250015,0.250008,0.375,0.25,0.1875,0.15625,0.140625,0.132812,0.128906,0.126953,0.125977,0.125488,0.125244,0.125122,0.125061,0.125031,0.125015,0.125008,0.3125,0.1875,0.125,0.09375,0.078125,0.0703125,0.0664062,0.0644531,0.0634766,0.0629883,0.0627441,0.0626221,0.062561,0.0625305,0.0625153,0.0625076,0.28125,0.15625,0.09375,0.0625,0.046875,0.0390625,0.0351562,0.0332031,0.0322266,0.0317383,0.0314941,0.0313721,0.031311,0.0312805,0.0312653,0.0312576,0.265625,0.140625,0.078125,0.046875,0.03125,0.0234375,0.0195312,0.0175781,0.0166016,0.0161133,0.0158691,0.0157471,0.015686,0.0156555,0.0156403,0.0156326,0.257812,0.132812,0.0703125,0.0390625,0.0234375,0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,0.00782776,0.00782013,0.253906,0.128906,0.0664062,0.0351562,0.0195312,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.00391388,0.251953,0.126953,0.0644531,0.0332031,0.0175781,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.250977,0.125977,0.0634766,0.0322266,0.0166016,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.250488,0.125488,0.0629883,0.0317383,0.0161133,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.250244,0.125244,0.0627441,0.0314941,0.0158691,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.250122,0.125122,0.0626221,0.0313721,0.0157471,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.250061,0.125061,0.062561,0.031311,0.015686,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,0.250031,0.125031,0.0625305,0.0312805,0.0156555,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,0.250015,0.125015,0.0625153,0.0312653,0.0156403,0.00782776,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,0.250008,0.125008,0.0625076,0.0312576,0.0156326,0.00782013,0.00391388,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,},
-{0.25,0.1875,0.15625,0.140625,0.132812,0.128906,0.126953,0.125977,0.125488,0.125244,0.125122,0.125061,0.125031,0.125015,0.125008,0.125004,0.1875,0.125,0.09375,0.078125,0.0703125,0.0664062,0.0644531,0.0634766,0.0629883,0.0627441,0.0626221,0.062561,0.0625305,0.0625153,0.0625076,0.0625038,0.15625,0.09375,0.0625,0.046875,0.0390625,0.0351562,0.0332031,0.0322266,0.0317383,0.0314941,0.0313721,0.031311,0.0312805,0.0312653,0.0312576,0.0312538,0.140625,0.078125,0.046875,0.03125,0.0234375,0.0195312,0.0175781,0.0166016,0.0161133,0.0158691,0.0157471,0.015686,0.0156555,0.0156403,0.0156326,0.0156288,0.132812,0.0703125,0.0390625,0.0234375,0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,0.00782776,0.00782013,0.00781631,0.128906,0.0664062,0.0351562,0.0195312,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.00391388,0.00391006,0.126953,0.0644531,0.0332031,0.0175781,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.00195694,0.125977,0.0634766,0.0322266,0.0166016,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.125488,0.0629883,0.0317383,0.0161133,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.125244,0.0627441,0.0314941,0.0158691,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.125122,0.0626221,0.0313721,0.0157471,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.125061,0.062561,0.031311,0.015686,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,0.125031,0.0625305,0.0312805,0.0156555,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,0.125015,0.0625153,0.0312653,0.0156403,0.00782776,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,0.125008,0.0625076,0.0312576,0.0156326,0.00782013,0.00391388,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,0.125004,0.0625038,0.0312538,0.0156288,0.00781631,0.00391006,0.00195694,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,},
-{0.125,0.09375,0.078125,0.0703125,0.0664062,0.0644531,0.0634766,0.0629883,0.0627441,0.0626221,0.062561,0.0625305,0.0625153,0.0625076,0.0625038,0.0625019,0.09375,0.0625,0.046875,0.0390625,0.0351562,0.0332031,0.0322266,0.0317383,0.0314941,0.0313721,0.031311,0.0312805,0.0312653,0.0312576,0.0312538,0.0312519,0.078125,0.046875,0.03125,0.0234375,0.0195312,0.0175781,0.0166016,0.0161133,0.0158691,0.0157471,0.015686,0.0156555,0.0156403,0.0156326,0.0156288,0.0156269,0.0703125,0.0390625,0.0234375,0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,0.00782776,0.00782013,0.00781631,0.00781441,0.0664062,0.0351562,0.0195312,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.00391388,0.00391006,0.00390816,0.0644531,0.0332031,0.0175781,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.00195694,0.00195503,0.0634766,0.0322266,0.0166016,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.00097847,0.0629883,0.0317383,0.0161133,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.0627441,0.0314941,0.0158691,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.0626221,0.0313721,0.0157471,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.062561,0.031311,0.015686,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,0.0625305,0.0312805,0.0156555,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,0.0625153,0.0312653,0.0156403,0.00782776,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,0.0625076,0.0312576,0.0156326,0.00782013,0.00391388,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,0.0625038,0.0312538,0.0156288,0.00781631,0.00391006,0.00195694,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,0.0625019,0.0312519,0.0156269,0.00781441,0.00390816,0.00195503,0.00097847,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,},
-{0.0625,0.046875,0.0390625,0.0351562,0.0332031,0.0322266,0.0317383,0.0314941,0.0313721,0.031311,0.0312805,0.0312653,0.0312576,0.0312538,0.0312519,0.031251,0.046875,0.03125,0.0234375,0.0195312,0.0175781,0.0166016,0.0161133,0.0158691,0.0157471,0.015686,0.0156555,0.0156403,0.0156326,0.0156288,0.0156269,0.015626,0.0390625,0.0234375,0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,0.00782776,0.00782013,0.00781631,0.00781441,0.00781345,0.0351562,0.0195312,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.00391388,0.00391006,0.00390816,0.0039072,0.0332031,0.0175781,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.00195694,0.00195503,0.00195408,0.0322266,0.0166016,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.00097847,0.000977516,0.0317383,0.0161133,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.000489235,0.0314941,0.0158691,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.0313721,0.0157471,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.031311,0.015686,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,0.0312805,0.0156555,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,0.0312653,0.0156403,0.00782776,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,0.0312576,0.0156326,0.00782013,0.00391388,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,0.0312538,0.0156288,0.00781631,0.00391006,0.00195694,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,0.0312519,0.0156269,0.00781441,0.00390816,0.00195503,0.00097847,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,0.031251,0.015626,0.00781345,0.0039072,0.00195408,0.000977516,0.000489235,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,},
-{0.03125,0.0234375,0.0195312,0.0175781,0.0166016,0.0161133,0.0158691,0.0157471,0.015686,0.0156555,0.0156403,0.0156326,0.0156288,0.0156269,0.015626,0.0156255,0.0234375,0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,0.00782776,0.00782013,0.00781631,0.00781441,0.00781345,0.00781298,0.0195312,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.00391388,0.00391006,0.00390816,0.0039072,0.00390673,0.0175781,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.00195694,0.00195503,0.00195408,0.0019536,0.0166016,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.00097847,0.000977516,0.000977039,0.0161133,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.000489235,0.000488758,0.0158691,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.000244617,0.0157471,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.015686,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,0.0156555,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,0.0156403,0.00782776,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,0.0156326,0.00782013,0.00391388,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,0.0156288,0.00781631,0.00391006,0.00195694,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,0.0156269,0.00781441,0.00390816,0.00195503,0.00097847,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,0.015626,0.00781345,0.0039072,0.00195408,0.000977516,0.000489235,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,0.0156255,0.00781298,0.00390673,0.0019536,0.000977039,0.000488758,0.000244617,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,},
-{0.015625,0.0117188,0.00976562,0.00878906,0.00830078,0.00805664,0.00793457,0.00787354,0.00784302,0.00782776,0.00782013,0.00781631,0.00781441,0.00781345,0.00781298,0.00781274,0.0117188,0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.00391388,0.00391006,0.00390816,0.0039072,0.00390673,0.00390649,0.00976562,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.00195694,0.00195503,0.00195408,0.0019536,0.00195336,0.00878906,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.00097847,0.000977516,0.000977039,0.000976801,0.00830078,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.000489235,0.000488758,0.00048852,0.00805664,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.000244617,0.000244379,0.00793457,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.000122309,0.00787354,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,0.00784302,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,0.00782776,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,0.00782013,0.00391388,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,0.00781631,0.00391006,0.00195694,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,0.00781441,0.00390816,0.00195503,0.00097847,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,0.00781345,0.0039072,0.00195408,0.000977516,0.000489235,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,0.00781298,0.00390673,0.0019536,0.000977039,0.000488758,0.000244617,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,0.00781274,0.00390649,0.00195336,0.000976801,0.00048852,0.000244379,0.000122309,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,},
-{0.0078125,0.00585938,0.00488281,0.00439453,0.00415039,0.00402832,0.00396729,0.00393677,0.00392151,0.00391388,0.00391006,0.00390816,0.0039072,0.00390673,0.00390649,0.00390637,0.00585938,0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.00195694,0.00195503,0.00195408,0.0019536,0.00195336,0.00195324,0.00488281,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.00097847,0.000977516,0.000977039,0.000976801,0.000976682,0.00439453,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.000489235,0.000488758,0.00048852,0.0004884,0.00415039,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.000244617,0.000244379,0.00024426,0.00402832,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.000122309,0.00012219,0.00396729,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,6.11544e-05,0.00393677,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,0.00392151,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,0.00391388,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,0.00391006,0.00195694,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,0.00390816,0.00195503,0.00097847,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,0.0039072,0.00195408,0.000977516,0.000489235,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,0.00390673,0.0019536,0.000977039,0.000488758,0.000244617,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,0.00390649,0.00195336,0.000976801,0.00048852,0.000244379,0.000122309,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,0.00390637,0.00195324,0.000976682,0.0004884,0.00024426,0.00012219,6.11544e-05,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,},
-{0.00390625,0.00292969,0.00244141,0.00219727,0.0020752,0.00201416,0.00198364,0.00196838,0.00196075,0.00195694,0.00195503,0.00195408,0.0019536,0.00195336,0.00195324,0.00195318,0.00292969,0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.00097847,0.000977516,0.000977039,0.000976801,0.000976682,0.000976622,0.00244141,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.000489235,0.000488758,0.00048852,0.0004884,0.000488341,0.00219727,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.000244617,0.000244379,0.00024426,0.0002442,0.0020752,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.000122309,0.00012219,0.00012213,0.00201416,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,6.11544e-05,6.10948e-05,0.00198364,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,3.05772e-05,0.00196838,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,0.00196075,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,0.00195694,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,0.00195503,0.00097847,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,0.00195408,0.000977516,0.000489235,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,0.0019536,0.000977039,0.000488758,0.000244617,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,0.00195336,0.000976801,0.00048852,0.000244379,0.000122309,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,0.00195324,0.000976682,0.0004884,0.00024426,0.00012219,6.11544e-05,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,0.00195318,0.000976622,0.000488341,0.0002442,0.00012213,6.10948e-05,3.05772e-05,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,},
-{0.00195312,0.00146484,0.0012207,0.00109863,0.0010376,0.00100708,0.000991821,0.000984192,0.000980377,0.00097847,0.000977516,0.000977039,0.000976801,0.000976682,0.000976622,0.000976592,0.00146484,0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.000489235,0.000488758,0.00048852,0.0004884,0.000488341,0.000488311,0.0012207,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.000244617,0.000244379,0.00024426,0.0002442,0.00024417,0.00109863,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.000122309,0.00012219,0.00012213,0.0001221,0.0010376,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,6.11544e-05,6.10948e-05,6.1065e-05,0.00100708,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,3.05772e-05,3.05474e-05,0.000991821,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,1.52886e-05,0.000984192,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,0.000980377,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,0.00097847,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,0.000977516,0.000489235,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,0.000977039,0.000488758,0.000244617,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,0.000976801,0.00048852,0.000244379,0.000122309,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,0.000976682,0.0004884,0.00024426,0.00012219,6.11544e-05,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,0.000976622,0.000488341,0.0002442,0.00012213,6.10948e-05,3.05772e-05,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,0.000976592,0.000488311,0.00024417,0.0001221,6.1065e-05,3.05474e-05,1.52886e-05,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,},
-{0.000976562,0.000732422,0.000610352,0.000549316,0.000518799,0.00050354,0.000495911,0.000492096,0.000490189,0.000489235,0.000488758,0.00048852,0.0004884,0.000488341,0.000488311,0.000488296,0.000732422,0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.000244617,0.000244379,0.00024426,0.0002442,0.00024417,0.000244156,0.000610352,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.000122309,0.00012219,0.00012213,0.0001221,0.000122085,0.000549316,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,6.11544e-05,6.10948e-05,6.1065e-05,6.10501e-05,0.000518799,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,3.05772e-05,3.05474e-05,3.05325e-05,0.00050354,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,1.52886e-05,1.52737e-05,0.000495911,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,7.6443e-06,0.000492096,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,0.000490189,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,0.000489235,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,0.000488758,0.000244617,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,0.00048852,0.000244379,0.000122309,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,0.0004884,0.00024426,0.00012219,6.11544e-05,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,0.000488341,0.0002442,0.00012213,6.10948e-05,3.05772e-05,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,0.000488311,0.00024417,0.0001221,6.1065e-05,3.05474e-05,1.52886e-05,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,0.000488296,0.000244156,0.000122085,6.10501e-05,3.05325e-05,1.52737e-05,7.6443e-06,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,},
-{0.000488281,0.000366211,0.000305176,0.000274658,0.000259399,0.00025177,0.000247955,0.000246048,0.000245094,0.000244617,0.000244379,0.00024426,0.0002442,0.00024417,0.000244156,0.000244148,0.000366211,0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.000122309,0.00012219,0.00012213,0.0001221,0.000122085,0.000122078,0.000305176,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,6.11544e-05,6.10948e-05,6.1065e-05,6.10501e-05,6.10426e-05,0.000274658,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,3.05772e-05,3.05474e-05,3.05325e-05,3.0525e-05,0.000259399,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,1.52886e-05,1.52737e-05,1.52662e-05,0.00025177,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,7.6443e-06,7.63685e-06,0.000247955,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,3.82215e-06,0.000246048,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,0.000245094,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,0.000244617,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,0.000244379,0.000122309,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,0.00024426,0.00012219,6.11544e-05,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,0.0002442,0.00012213,6.10948e-05,3.05772e-05,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,0.00024417,0.0001221,6.1065e-05,3.05474e-05,1.52886e-05,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,0.000244156,0.000122085,6.10501e-05,3.05325e-05,1.52737e-05,7.6443e-06,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,0.000244148,0.000122078,6.10426e-05,3.0525e-05,1.52662e-05,7.63685e-06,3.82215e-06,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,},
-{0.000244141,0.000183105,0.000152588,0.000137329,0.0001297,0.000125885,0.000123978,0.000123024,0.000122547,0.000122309,0.00012219,0.00012213,0.0001221,0.000122085,0.000122078,0.000122074,0.000183105,0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,6.11544e-05,6.10948e-05,6.1065e-05,6.10501e-05,6.10426e-05,6.10389e-05,0.000152588,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,3.05772e-05,3.05474e-05,3.05325e-05,3.0525e-05,3.05213e-05,0.000137329,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,1.52886e-05,1.52737e-05,1.52662e-05,1.52625e-05,0.0001297,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,7.6443e-06,7.63685e-06,7.63312e-06,0.000125885,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,3.82215e-06,3.81842e-06,0.000123978,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,1.91107e-06,0.000123024,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,0.000122547,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,0.000122309,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,0.00012219,6.11544e-05,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,0.00012213,6.10948e-05,3.05772e-05,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,0.0001221,6.1065e-05,3.05474e-05,1.52886e-05,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,0.000122085,6.10501e-05,3.05325e-05,1.52737e-05,7.6443e-06,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,0.000122078,6.10426e-05,3.0525e-05,1.52662e-05,7.63685e-06,3.82215e-06,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,0.000122074,6.10389e-05,3.05213e-05,1.52625e-05,7.63312e-06,3.81842e-06,1.91107e-06,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,},
-{0.00012207,9.15527e-05,7.62939e-05,6.86646e-05,6.48499e-05,6.29425e-05,6.19888e-05,6.1512e-05,6.12736e-05,6.11544e-05,6.10948e-05,6.1065e-05,6.10501e-05,6.10426e-05,6.10389e-05,6.1037e-05,9.15527e-05,6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,3.05772e-05,3.05474e-05,3.05325e-05,3.0525e-05,3.05213e-05,3.05194e-05,7.62939e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,1.52886e-05,1.52737e-05,1.52662e-05,1.52625e-05,1.52607e-05,6.86646e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,7.6443e-06,7.63685e-06,7.63312e-06,7.63126e-06,6.48499e-05,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,3.82215e-06,3.81842e-06,3.81656e-06,6.29425e-05,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,1.91107e-06,1.90921e-06,6.19888e-05,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,9.55537e-07,6.1512e-05,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,6.12736e-05,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,6.11544e-05,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,6.10948e-05,3.05772e-05,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.1065e-05,3.05474e-05,1.52886e-05,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,6.10501e-05,3.05325e-05,1.52737e-05,7.6443e-06,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,6.10426e-05,3.0525e-05,1.52662e-05,7.63685e-06,3.82215e-06,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,6.10389e-05,3.05213e-05,1.52625e-05,7.63312e-06,3.81842e-06,1.91107e-06,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,6.1037e-05,3.05194e-05,1.52607e-05,7.63126e-06,3.81656e-06,1.90921e-06,9.55537e-07,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,},
-{6.10352e-05,4.57764e-05,3.8147e-05,3.43323e-05,3.24249e-05,3.14713e-05,3.09944e-05,3.0756e-05,3.06368e-05,3.05772e-05,3.05474e-05,3.05325e-05,3.0525e-05,3.05213e-05,3.05194e-05,3.05185e-05,4.57764e-05,3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,1.52886e-05,1.52737e-05,1.52662e-05,1.52625e-05,1.52607e-05,1.52597e-05,3.8147e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,7.6443e-06,7.63685e-06,7.63312e-06,7.63126e-06,7.63033e-06,3.43323e-05,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,3.82215e-06,3.81842e-06,3.81656e-06,3.81563e-06,3.24249e-05,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,1.91107e-06,1.90921e-06,1.90828e-06,3.14713e-05,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,9.55537e-07,9.54606e-07,3.09944e-05,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,4.77768e-07,3.0756e-05,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,3.06368e-05,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,3.05772e-05,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,3.05474e-05,1.52886e-05,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.05325e-05,1.52737e-05,7.6443e-06,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,3.0525e-05,1.52662e-05,7.63685e-06,3.82215e-06,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,3.05213e-05,1.52625e-05,7.63312e-06,3.81842e-06,1.91107e-06,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,3.05194e-05,1.52607e-05,7.63126e-06,3.81656e-06,1.90921e-06,9.55537e-07,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,3.05185e-05,1.52597e-05,7.63033e-06,3.81563e-06,1.90828e-06,9.54606e-07,4.77768e-07,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,},
-{3.05176e-05,2.28882e-05,1.90735e-05,1.71661e-05,1.62125e-05,1.57356e-05,1.54972e-05,1.5378e-05,1.53184e-05,1.52886e-05,1.52737e-05,1.52662e-05,1.52625e-05,1.52607e-05,1.52597e-05,1.52593e-05,2.28882e-05,1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,7.6443e-06,7.63685e-06,7.63312e-06,7.63126e-06,7.63033e-06,7.62986e-06,1.90735e-05,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,3.82215e-06,3.81842e-06,3.81656e-06,3.81563e-06,3.81516e-06,1.71661e-05,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,1.91107e-06,1.90921e-06,1.90828e-06,1.90781e-06,1.62125e-05,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,9.55537e-07,9.54606e-07,9.5414e-07,1.57356e-05,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,4.77768e-07,4.77303e-07,1.54972e-05,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,2.38884e-07,1.5378e-05,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.53184e-05,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,1.52886e-05,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,1.52737e-05,7.6443e-06,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.52662e-05,7.63685e-06,3.82215e-06,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,1.52625e-05,7.63312e-06,3.81842e-06,1.91107e-06,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,1.52607e-05,7.63126e-06,3.81656e-06,1.90921e-06,9.55537e-07,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,1.52597e-05,7.63033e-06,3.81563e-06,1.90828e-06,9.54606e-07,4.77768e-07,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.52593e-05,7.62986e-06,3.81516e-06,1.90781e-06,9.5414e-07,4.77303e-07,2.38884e-07,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,},
-{1.52588e-05,1.14441e-05,9.53674e-06,8.58307e-06,8.10623e-06,7.86781e-06,7.7486e-06,7.689e-06,7.6592e-06,7.6443e-06,7.63685e-06,7.63312e-06,7.63126e-06,7.63033e-06,7.62986e-06,7.62963e-06,1.14441e-05,7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,3.82215e-06,3.81842e-06,3.81656e-06,3.81563e-06,3.81516e-06,3.81493e-06,9.53674e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,1.91107e-06,1.90921e-06,1.90828e-06,1.90781e-06,1.90758e-06,8.58307e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,9.55537e-07,9.54606e-07,9.5414e-07,9.53907e-07,8.10623e-06,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,4.77768e-07,4.77303e-07,4.7707e-07,7.86781e-06,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,2.38884e-07,2.38651e-07,7.7486e-06,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.19442e-07,7.689e-06,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,7.6592e-06,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,7.6443e-06,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,7.63685e-06,3.82215e-06,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.63312e-06,3.81842e-06,1.91107e-06,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,7.63126e-06,3.81656e-06,1.90921e-06,9.55537e-07,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,7.63033e-06,3.81563e-06,1.90828e-06,9.54606e-07,4.77768e-07,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,7.62986e-06,3.81516e-06,1.90781e-06,9.5414e-07,4.77303e-07,2.38884e-07,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,7.62963e-06,3.81493e-06,1.90758e-06,9.53907e-07,4.7707e-07,2.38651e-07,1.19442e-07,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,},
-{7.62939e-06,5.72205e-06,4.76837e-06,4.29153e-06,4.05312e-06,3.93391e-06,3.8743e-06,3.8445e-06,3.8296e-06,3.82215e-06,3.81842e-06,3.81656e-06,3.81563e-06,3.81516e-06,3.81493e-06,3.81481e-06,5.72205e-06,3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,1.91107e-06,1.90921e-06,1.90828e-06,1.90781e-06,1.90758e-06,1.90747e-06,4.76837e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,9.55537e-07,9.54606e-07,9.5414e-07,9.53907e-07,9.53791e-07,4.29153e-06,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,4.77768e-07,4.77303e-07,4.7707e-07,4.76954e-07,4.05312e-06,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,2.38884e-07,2.38651e-07,2.38535e-07,3.93391e-06,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.19442e-07,1.19326e-07,3.8743e-06,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,5.97211e-08,3.8445e-06,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,3.8296e-06,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,3.82215e-06,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,3.81842e-06,1.91107e-06,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.81656e-06,1.90921e-06,9.55537e-07,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,3.81563e-06,1.90828e-06,9.54606e-07,4.77768e-07,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,3.81516e-06,1.90781e-06,9.5414e-07,4.77303e-07,2.38884e-07,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,3.81493e-06,1.90758e-06,9.53907e-07,4.7707e-07,2.38651e-07,1.19442e-07,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,3.81481e-06,1.90747e-06,9.53791e-07,4.76954e-07,2.38535e-07,1.19326e-07,5.97211e-08,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,},
-{3.8147e-06,2.86102e-06,2.38419e-06,2.14577e-06,2.02656e-06,1.96695e-06,1.93715e-06,1.92225e-06,1.9148e-06,1.91107e-06,1.90921e-06,1.90828e-06,1.90781e-06,1.90758e-06,1.90747e-06,1.90741e-06,2.86102e-06,1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,9.55537e-07,9.54606e-07,9.5414e-07,9.53907e-07,9.53791e-07,9.53733e-07,2.38419e-06,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,4.77768e-07,4.77303e-07,4.7707e-07,4.76954e-07,4.76895e-07,2.14577e-06,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,2.38884e-07,2.38651e-07,2.38535e-07,2.38477e-07,2.02656e-06,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.19442e-07,1.19326e-07,1.19267e-07,1.96695e-06,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,5.97211e-08,5.96629e-08,1.93715e-06,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,2.98605e-08,1.92225e-06,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.9148e-06,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,1.91107e-06,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,1.90921e-06,9.55537e-07,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.90828e-06,9.54606e-07,4.77768e-07,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,1.90781e-06,9.5414e-07,4.77303e-07,2.38884e-07,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,1.90758e-06,9.53907e-07,4.7707e-07,2.38651e-07,1.19442e-07,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,1.90747e-06,9.53791e-07,4.76954e-07,2.38535e-07,1.19326e-07,5.97211e-08,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.90741e-06,9.53733e-07,4.76895e-07,2.38477e-07,1.19267e-07,5.96629e-08,2.98605e-08,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,},
-{1.90735e-06,1.43051e-06,1.19209e-06,1.07288e-06,1.01328e-06,9.83477e-07,9.68575e-07,9.61125e-07,9.574e-07,9.55537e-07,9.54606e-07,9.5414e-07,9.53907e-07,9.53791e-07,9.53733e-07,9.53703e-07,1.43051e-06,9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,4.77768e-07,4.77303e-07,4.7707e-07,4.76954e-07,4.76895e-07,4.76866e-07,1.19209e-06,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,2.38884e-07,2.38651e-07,2.38535e-07,2.38477e-07,2.38448e-07,1.07288e-06,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.19442e-07,1.19326e-07,1.19267e-07,1.19238e-07,1.01328e-06,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,5.97211e-08,5.96629e-08,5.96337e-08,9.83477e-07,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,2.98605e-08,2.98314e-08,9.68575e-07,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.49303e-08,9.61125e-07,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,9.574e-07,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,9.55537e-07,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,9.54606e-07,4.77768e-07,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.5414e-07,4.77303e-07,2.38884e-07,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,9.53907e-07,4.7707e-07,2.38651e-07,1.19442e-07,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,9.53791e-07,4.76954e-07,2.38535e-07,1.19326e-07,5.97211e-08,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,9.53733e-07,4.76895e-07,2.38477e-07,1.19267e-07,5.96629e-08,2.98605e-08,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,9.53703e-07,4.76866e-07,2.38448e-07,1.19238e-07,5.96337e-08,2.98314e-08,1.49303e-08,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,},
-{9.53674e-07,7.15256e-07,5.96046e-07,5.36442e-07,5.06639e-07,4.91738e-07,4.84288e-07,4.80562e-07,4.787e-07,4.77768e-07,4.77303e-07,4.7707e-07,4.76954e-07,4.76895e-07,4.76866e-07,4.76852e-07,7.15256e-07,4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,2.38884e-07,2.38651e-07,2.38535e-07,2.38477e-07,2.38448e-07,2.38433e-07,5.96046e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.19442e-07,1.19326e-07,1.19267e-07,1.19238e-07,1.19224e-07,5.36442e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,5.97211e-08,5.96629e-08,5.96337e-08,5.96192e-08,5.06639e-07,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,2.98605e-08,2.98314e-08,2.98169e-08,4.91738e-07,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.49303e-08,1.49157e-08,4.84288e-07,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,7.46513e-09,4.80562e-07,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,4.787e-07,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,4.77768e-07,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,4.77303e-07,2.38884e-07,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.7707e-07,2.38651e-07,1.19442e-07,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,4.76954e-07,2.38535e-07,1.19326e-07,5.97211e-08,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,4.76895e-07,2.38477e-07,1.19267e-07,5.96629e-08,2.98605e-08,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,4.76866e-07,2.38448e-07,1.19238e-07,5.96337e-08,2.98314e-08,1.49303e-08,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,4.76852e-07,2.38433e-07,1.19224e-07,5.96192e-08,2.98169e-08,1.49157e-08,7.46513e-09,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,},
-{4.76837e-07,3.57628e-07,2.98023e-07,2.68221e-07,2.5332e-07,2.45869e-07,2.42144e-07,2.40281e-07,2.3935e-07,2.38884e-07,2.38651e-07,2.38535e-07,2.38477e-07,2.38448e-07,2.38433e-07,2.38426e-07,3.57628e-07,2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.19442e-07,1.19326e-07,1.19267e-07,1.19238e-07,1.19224e-07,1.19217e-07,2.98023e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,5.97211e-08,5.96629e-08,5.96337e-08,5.96192e-08,5.96119e-08,2.68221e-07,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,2.98605e-08,2.98314e-08,2.98169e-08,2.98096e-08,2.5332e-07,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.49303e-08,1.49157e-08,1.49084e-08,2.45869e-07,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,7.46513e-09,7.45786e-09,2.42144e-07,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,3.73257e-09,2.40281e-07,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,2.3935e-07,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,2.38884e-07,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,2.38651e-07,1.19442e-07,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.38535e-07,1.19326e-07,5.97211e-08,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,2.38477e-07,1.19267e-07,5.96629e-08,2.98605e-08,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,2.38448e-07,1.19238e-07,5.96337e-08,2.98314e-08,1.49303e-08,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,2.38433e-07,1.19224e-07,5.96192e-08,2.98169e-08,1.49157e-08,7.46513e-09,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,2.38426e-07,1.19217e-07,5.96119e-08,2.98096e-08,1.49084e-08,7.45786e-09,3.73257e-09,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,},
-{2.38419e-07,1.78814e-07,1.49012e-07,1.3411e-07,1.2666e-07,1.22935e-07,1.21072e-07,1.20141e-07,1.19675e-07,1.19442e-07,1.19326e-07,1.19267e-07,1.19238e-07,1.19224e-07,1.19217e-07,1.19213e-07,1.78814e-07,1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,5.97211e-08,5.96629e-08,5.96337e-08,5.96192e-08,5.96119e-08,5.96083e-08,1.49012e-07,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,2.98605e-08,2.98314e-08,2.98169e-08,2.98096e-08,2.9806e-08,1.3411e-07,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.49303e-08,1.49157e-08,1.49084e-08,1.49048e-08,1.2666e-07,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,7.46513e-09,7.45786e-09,7.45422e-09,1.22935e-07,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,3.73257e-09,3.72893e-09,1.21072e-07,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,1.86628e-09,1.20141e-07,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,1.19675e-07,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,1.19442e-07,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,1.19326e-07,5.97211e-08,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.19267e-07,5.96629e-08,2.98605e-08,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,1.19238e-07,5.96337e-08,2.98314e-08,1.49303e-08,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,1.19224e-07,5.96192e-08,2.98169e-08,1.49157e-08,7.46513e-09,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.19217e-07,5.96119e-08,2.98096e-08,1.49084e-08,7.45786e-09,3.73257e-09,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,1.19213e-07,5.96083e-08,2.9806e-08,1.49048e-08,7.45422e-09,3.72893e-09,1.86628e-09,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,},
-{1.19209e-07,8.9407e-08,7.45058e-08,6.70552e-08,6.33299e-08,6.14673e-08,6.0536e-08,6.00703e-08,5.98375e-08,5.97211e-08,5.96629e-08,5.96337e-08,5.96192e-08,5.96119e-08,5.96083e-08,5.96065e-08,8.9407e-08,5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,2.98605e-08,2.98314e-08,2.98169e-08,2.98096e-08,2.9806e-08,2.98041e-08,7.45058e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.49303e-08,1.49157e-08,1.49084e-08,1.49048e-08,1.4903e-08,6.70552e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,7.46513e-09,7.45786e-09,7.45422e-09,7.4524e-09,6.33299e-08,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,3.73257e-09,3.72893e-09,3.72711e-09,6.14673e-08,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,1.86628e-09,1.86446e-09,6.0536e-08,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,9.33142e-10,6.00703e-08,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,5.98375e-08,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,5.97211e-08,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,5.96629e-08,2.98605e-08,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.96337e-08,2.98314e-08,1.49303e-08,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,5.96192e-08,2.98169e-08,1.49157e-08,7.46513e-09,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,5.96119e-08,2.98096e-08,1.49084e-08,7.45786e-09,3.73257e-09,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,5.96083e-08,2.9806e-08,1.49048e-08,7.45422e-09,3.72893e-09,1.86628e-09,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,5.96065e-08,2.98041e-08,1.4903e-08,7.4524e-09,3.72711e-09,1.86446e-09,9.33142e-10,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,},
-{5.96046e-08,4.47035e-08,3.72529e-08,3.35276e-08,3.1665e-08,3.07336e-08,3.0268e-08,3.00352e-08,2.99187e-08,2.98605e-08,2.98314e-08,2.98169e-08,2.98096e-08,2.9806e-08,2.98041e-08,2.98032e-08,4.47035e-08,2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.49303e-08,1.49157e-08,1.49084e-08,1.49048e-08,1.4903e-08,1.49021e-08,3.72529e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,7.46513e-09,7.45786e-09,7.45422e-09,7.4524e-09,7.45149e-09,3.35276e-08,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,3.73257e-09,3.72893e-09,3.72711e-09,3.7262e-09,3.1665e-08,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,1.86628e-09,1.86446e-09,1.86355e-09,3.07336e-08,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,9.33142e-10,9.32232e-10,3.0268e-08,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,4.66571e-10,3.00352e-08,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.99187e-08,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,2.98605e-08,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,2.98314e-08,1.49303e-08,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.98169e-08,1.49157e-08,7.46513e-09,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,2.98096e-08,1.49084e-08,7.45786e-09,3.73257e-09,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,2.9806e-08,1.49048e-08,7.45422e-09,3.72893e-09,1.86628e-09,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,2.98041e-08,1.4903e-08,7.4524e-09,3.72711e-09,1.86446e-09,9.33142e-10,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.98032e-08,1.49021e-08,7.45149e-09,3.7262e-09,1.86355e-09,9.32232e-10,4.66571e-10,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,},
-{2.98023e-08,2.23517e-08,1.86265e-08,1.67638e-08,1.58325e-08,1.53668e-08,1.5134e-08,1.50176e-08,1.49594e-08,1.49303e-08,1.49157e-08,1.49084e-08,1.49048e-08,1.4903e-08,1.49021e-08,1.49016e-08,2.23517e-08,1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,7.46513e-09,7.45786e-09,7.45422e-09,7.4524e-09,7.45149e-09,7.45104e-09,1.86265e-08,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,3.73257e-09,3.72893e-09,3.72711e-09,3.7262e-09,3.72575e-09,1.67638e-08,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,1.86628e-09,1.86446e-09,1.86355e-09,1.8631e-09,1.58325e-08,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,9.33142e-10,9.32232e-10,9.31777e-10,1.53668e-08,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,4.66571e-10,4.66116e-10,1.5134e-08,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.33285e-10,1.50176e-08,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.49594e-08,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,1.49303e-08,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,1.49157e-08,7.46513e-09,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.49084e-08,7.45786e-09,3.73257e-09,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,1.49048e-08,7.45422e-09,3.72893e-09,1.86628e-09,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,1.4903e-08,7.4524e-09,3.72711e-09,1.86446e-09,9.33142e-10,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,1.49021e-08,7.45149e-09,3.7262e-09,1.86355e-09,9.32232e-10,4.66571e-10,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.49016e-08,7.45104e-09,3.72575e-09,1.8631e-09,9.31777e-10,4.66116e-10,2.33285e-10,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,},
-{1.49012e-08,1.11759e-08,9.31323e-09,8.3819e-09,7.91624e-09,7.68341e-09,7.567e-09,7.50879e-09,7.47968e-09,7.46513e-09,7.45786e-09,7.45422e-09,7.4524e-09,7.45149e-09,7.45104e-09,7.45081e-09,1.11759e-08,7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,3.73257e-09,3.72893e-09,3.72711e-09,3.7262e-09,3.72575e-09,3.72552e-09,9.31323e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,1.86628e-09,1.86446e-09,1.86355e-09,1.8631e-09,1.86287e-09,8.3819e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,9.33142e-10,9.32232e-10,9.31777e-10,9.3155e-10,7.91624e-09,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,4.66571e-10,4.66116e-10,4.65889e-10,7.68341e-09,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.33285e-10,2.33058e-10,7.567e-09,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.16643e-10,7.50879e-09,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,7.47968e-09,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,7.46513e-09,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,7.45786e-09,3.73257e-09,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.45422e-09,3.72893e-09,1.86628e-09,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,7.4524e-09,3.72711e-09,1.86446e-09,9.33142e-10,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,7.45149e-09,3.7262e-09,1.86355e-09,9.32232e-10,4.66571e-10,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,7.45104e-09,3.72575e-09,1.8631e-09,9.31777e-10,4.66116e-10,2.33285e-10,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,7.45081e-09,3.72552e-09,1.86287e-09,9.3155e-10,4.65889e-10,2.33058e-10,1.16643e-10,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,},
-{7.45058e-09,5.58794e-09,4.65661e-09,4.19095e-09,3.95812e-09,3.84171e-09,3.7835e-09,3.75439e-09,3.73984e-09,3.73257e-09,3.72893e-09,3.72711e-09,3.7262e-09,3.72575e-09,3.72552e-09,3.7254e-09,5.58794e-09,3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,1.86628e-09,1.86446e-09,1.86355e-09,1.8631e-09,1.86287e-09,1.86276e-09,4.65661e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,9.33142e-10,9.32232e-10,9.31777e-10,9.3155e-10,9.31436e-10,4.19095e-09,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,4.66571e-10,4.66116e-10,4.65889e-10,4.65775e-10,3.95812e-09,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.33285e-10,2.33058e-10,2.32944e-10,3.84171e-09,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.16643e-10,1.16529e-10,3.7835e-09,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,5.83213e-11,3.75439e-09,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,3.73984e-09,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,3.73257e-09,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,3.72893e-09,1.86628e-09,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.72711e-09,1.86446e-09,9.33142e-10,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,3.7262e-09,1.86355e-09,9.32232e-10,4.66571e-10,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,3.72575e-09,1.8631e-09,9.31777e-10,4.66116e-10,2.33285e-10,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,3.72552e-09,1.86287e-09,9.3155e-10,4.65889e-10,2.33058e-10,1.16643e-10,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,3.7254e-09,1.86276e-09,9.31436e-10,4.65775e-10,2.32944e-10,1.16529e-10,5.83213e-11,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,},
-{3.72529e-09,2.79397e-09,2.32831e-09,2.09548e-09,1.97906e-09,1.92085e-09,1.89175e-09,1.8772e-09,1.86992e-09,1.86628e-09,1.86446e-09,1.86355e-09,1.8631e-09,1.86287e-09,1.86276e-09,1.8627e-09,2.79397e-09,1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,9.33142e-10,9.32232e-10,9.31777e-10,9.3155e-10,9.31436e-10,9.31379e-10,2.32831e-09,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,4.66571e-10,4.66116e-10,4.65889e-10,4.65775e-10,4.65718e-10,2.09548e-09,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.33285e-10,2.33058e-10,2.32944e-10,2.32887e-10,1.97906e-09,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.16643e-10,1.16529e-10,1.16472e-10,1.92085e-09,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,5.83213e-11,5.82645e-11,1.89175e-09,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,2.91607e-11,1.8772e-09,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.86992e-09,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,1.86628e-09,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,1.86446e-09,9.33142e-10,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.86355e-09,9.32232e-10,4.66571e-10,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,1.8631e-09,9.31777e-10,4.66116e-10,2.33285e-10,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,1.86287e-09,9.3155e-10,4.65889e-10,2.33058e-10,1.16643e-10,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,1.86276e-09,9.31436e-10,4.65775e-10,2.32944e-10,1.16529e-10,5.83213e-11,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.8627e-09,9.31379e-10,4.65718e-10,2.32887e-10,1.16472e-10,5.82645e-11,2.91607e-11,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,},
-{1.86265e-09,1.39698e-09,1.16415e-09,1.04774e-09,9.8953e-10,9.60426e-10,9.45874e-10,9.38599e-10,9.34961e-10,9.33142e-10,9.32232e-10,9.31777e-10,9.3155e-10,9.31436e-10,9.31379e-10,9.31351e-10,1.39698e-09,9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,4.66571e-10,4.66116e-10,4.65889e-10,4.65775e-10,4.65718e-10,4.6569e-10,1.16415e-09,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.33285e-10,2.33058e-10,2.32944e-10,2.32887e-10,2.32859e-10,1.04774e-09,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.16643e-10,1.16529e-10,1.16472e-10,1.16444e-10,9.8953e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,5.83213e-11,5.82645e-11,5.82361e-11,9.60426e-10,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,2.91607e-11,2.91323e-11,9.45874e-10,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.45803e-11,9.38599e-10,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,9.34961e-10,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,9.33142e-10,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,9.32232e-10,4.66571e-10,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.31777e-10,4.66116e-10,2.33285e-10,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,9.3155e-10,4.65889e-10,2.33058e-10,1.16643e-10,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,9.31436e-10,4.65775e-10,2.32944e-10,1.16529e-10,5.83213e-11,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,9.31379e-10,4.65718e-10,2.32887e-10,1.16472e-10,5.82645e-11,2.91607e-11,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,9.31351e-10,4.6569e-10,2.32859e-10,1.16444e-10,5.82361e-11,2.91323e-11,1.45803e-11,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,},
-{9.31323e-10,6.98492e-10,5.82077e-10,5.23869e-10,4.94765e-10,4.80213e-10,4.72937e-10,4.69299e-10,4.6748e-10,4.66571e-10,4.66116e-10,4.65889e-10,4.65775e-10,4.65718e-10,4.6569e-10,4.65675e-10,6.98492e-10,4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.33285e-10,2.33058e-10,2.32944e-10,2.32887e-10,2.32859e-10,2.32845e-10,5.82077e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.16643e-10,1.16529e-10,1.16472e-10,1.16444e-10,1.1643e-10,5.23869e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,5.83213e-11,5.82645e-11,5.82361e-11,5.82219e-11,4.94765e-10,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,2.91607e-11,2.91323e-11,2.9118e-11,4.80213e-10,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.45803e-11,1.45661e-11,4.72937e-10,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,7.29017e-12,4.69299e-10,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,4.6748e-10,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,4.66571e-10,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,4.66116e-10,2.33285e-10,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.65889e-10,2.33058e-10,1.16643e-10,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,4.65775e-10,2.32944e-10,1.16529e-10,5.83213e-11,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,4.65718e-10,2.32887e-10,1.16472e-10,5.82645e-11,2.91607e-11,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,4.6569e-10,2.32859e-10,1.16444e-10,5.82361e-11,2.91323e-11,1.45803e-11,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,4.65675e-10,2.32845e-10,1.1643e-10,5.82219e-11,2.9118e-11,1.45661e-11,7.29017e-12,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,},
-{4.65661e-10,3.49246e-10,2.91038e-10,2.61934e-10,2.47383e-10,2.40107e-10,2.36469e-10,2.3465e-10,2.3374e-10,2.33285e-10,2.33058e-10,2.32944e-10,2.32887e-10,2.32859e-10,2.32845e-10,2.32838e-10,3.49246e-10,2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.16643e-10,1.16529e-10,1.16472e-10,1.16444e-10,1.1643e-10,1.16422e-10,2.91038e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,5.83213e-11,5.82645e-11,5.82361e-11,5.82219e-11,5.82148e-11,2.61934e-10,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,2.91607e-11,2.91323e-11,2.9118e-11,2.91109e-11,2.47383e-10,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.45803e-11,1.45661e-11,1.4559e-11,2.40107e-10,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,7.29017e-12,7.28306e-12,2.36469e-10,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,3.64508e-12,2.3465e-10,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,2.3374e-10,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,2.33285e-10,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,2.33058e-10,1.16643e-10,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.32944e-10,1.16529e-10,5.83213e-11,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,2.32887e-10,1.16472e-10,5.82645e-11,2.91607e-11,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,2.32859e-10,1.16444e-10,5.82361e-11,2.91323e-11,1.45803e-11,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,2.32845e-10,1.1643e-10,5.82219e-11,2.9118e-11,1.45661e-11,7.29017e-12,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,2.32838e-10,1.16422e-10,5.82148e-11,2.91109e-11,1.4559e-11,7.28306e-12,3.64508e-12,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,},
-{2.32831e-10,1.74623e-10,1.45519e-10,1.30967e-10,1.23691e-10,1.20053e-10,1.18234e-10,1.17325e-10,1.1687e-10,1.16643e-10,1.16529e-10,1.16472e-10,1.16444e-10,1.1643e-10,1.16422e-10,1.16419e-10,1.74623e-10,1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,5.83213e-11,5.82645e-11,5.82361e-11,5.82219e-11,5.82148e-11,5.82112e-11,1.45519e-10,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,2.91607e-11,2.91323e-11,2.9118e-11,2.91109e-11,2.91074e-11,1.30967e-10,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.45803e-11,1.45661e-11,1.4559e-11,1.45555e-11,1.23691e-10,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,7.29017e-12,7.28306e-12,7.27951e-12,1.20053e-10,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,3.64508e-12,3.64153e-12,1.18234e-10,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,1.82254e-12,1.17325e-10,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,1.1687e-10,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,1.16643e-10,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,1.16529e-10,5.83213e-11,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.16472e-10,5.82645e-11,2.91607e-11,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,1.16444e-10,5.82361e-11,2.91323e-11,1.45803e-11,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,1.1643e-10,5.82219e-11,2.9118e-11,1.45661e-11,7.29017e-12,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.16422e-10,5.82148e-11,2.91109e-11,1.4559e-11,7.28306e-12,3.64508e-12,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,1.16419e-10,5.82112e-11,2.91074e-11,1.45555e-11,7.27951e-12,3.64153e-12,1.82254e-12,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,},
-{1.16415e-10,8.73115e-11,7.27596e-11,6.54836e-11,6.18456e-11,6.00267e-11,5.91172e-11,5.86624e-11,5.8435e-11,5.83213e-11,5.82645e-11,5.82361e-11,5.82219e-11,5.82148e-11,5.82112e-11,5.82094e-11,8.73115e-11,5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,2.91607e-11,2.91323e-11,2.9118e-11,2.91109e-11,2.91074e-11,2.91056e-11,7.27596e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.45803e-11,1.45661e-11,1.4559e-11,1.45555e-11,1.45537e-11,6.54836e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,7.29017e-12,7.28306e-12,7.27951e-12,7.27773e-12,6.18456e-11,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,3.64508e-12,3.64153e-12,3.63976e-12,6.00267e-11,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,1.82254e-12,1.82077e-12,5.91172e-11,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,9.11271e-13,5.86624e-11,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,5.8435e-11,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,5.83213e-11,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,5.82645e-11,2.91607e-11,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.82361e-11,2.91323e-11,1.45803e-11,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,5.82219e-11,2.9118e-11,1.45661e-11,7.29017e-12,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,5.82148e-11,2.91109e-11,1.4559e-11,7.28306e-12,3.64508e-12,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,5.82112e-11,2.91074e-11,1.45555e-11,7.27951e-12,3.64153e-12,1.82254e-12,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,5.82094e-11,2.91056e-11,1.45537e-11,7.27773e-12,3.63976e-12,1.82077e-12,9.11271e-13,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,},
-{5.82077e-11,4.36557e-11,3.63798e-11,3.27418e-11,3.09228e-11,3.00133e-11,2.95586e-11,2.93312e-11,2.92175e-11,2.91607e-11,2.91323e-11,2.9118e-11,2.91109e-11,2.91074e-11,2.91056e-11,2.91047e-11,4.36557e-11,2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.45803e-11,1.45661e-11,1.4559e-11,1.45555e-11,1.45537e-11,1.45528e-11,3.63798e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,7.29017e-12,7.28306e-12,7.27951e-12,7.27773e-12,7.27685e-12,3.27418e-11,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,3.64508e-12,3.64153e-12,3.63976e-12,3.63887e-12,3.09228e-11,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,1.82254e-12,1.82077e-12,1.81988e-12,3.00133e-11,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,9.11271e-13,9.10383e-13,2.95586e-11,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,4.55636e-13,2.93312e-11,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.92175e-11,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,2.91607e-11,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,2.91323e-11,1.45803e-11,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.9118e-11,1.45661e-11,7.29017e-12,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,2.91109e-11,1.4559e-11,7.28306e-12,3.64508e-12,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,2.91074e-11,1.45555e-11,7.27951e-12,3.64153e-12,1.82254e-12,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,2.91056e-11,1.45537e-11,7.27773e-12,3.63976e-12,1.82077e-12,9.11271e-13,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.91047e-11,1.45528e-11,7.27685e-12,3.63887e-12,1.81988e-12,9.10383e-13,4.55636e-13,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,},
-{2.91038e-11,2.18279e-11,1.81899e-11,1.63709e-11,1.54614e-11,1.50067e-11,1.47793e-11,1.46656e-11,1.46088e-11,1.45803e-11,1.45661e-11,1.4559e-11,1.45555e-11,1.45537e-11,1.45528e-11,1.45524e-11,2.18279e-11,1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,7.29017e-12,7.28306e-12,7.27951e-12,7.27773e-12,7.27685e-12,7.2764e-12,1.81899e-11,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,3.64508e-12,3.64153e-12,3.63976e-12,3.63887e-12,3.63842e-12,1.63709e-11,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,1.82254e-12,1.82077e-12,1.81988e-12,1.81943e-12,1.54614e-11,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,9.11271e-13,9.10383e-13,9.09939e-13,1.50067e-11,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,4.55636e-13,4.55191e-13,1.47793e-11,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.27818e-13,1.46656e-11,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.46088e-11,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,1.45803e-11,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,1.45661e-11,7.29017e-12,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.4559e-11,7.28306e-12,3.64508e-12,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,1.45555e-11,7.27951e-12,3.64153e-12,1.82254e-12,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,1.45537e-11,7.27773e-12,3.63976e-12,1.82077e-12,9.11271e-13,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.45528e-11,7.27685e-12,3.63887e-12,1.81988e-12,9.10383e-13,4.55636e-13,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.45524e-11,7.2764e-12,3.63842e-12,1.81943e-12,9.09939e-13,4.55191e-13,2.27818e-13,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,},
-{1.45519e-11,1.09139e-11,9.09495e-12,8.18545e-12,7.7307e-12,7.50333e-12,7.38964e-12,7.3328e-12,7.30438e-12,7.29017e-12,7.28306e-12,7.27951e-12,7.27773e-12,7.27685e-12,7.2764e-12,7.27618e-12,1.09139e-11,7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,3.64508e-12,3.64153e-12,3.63976e-12,3.63887e-12,3.63842e-12,3.6382e-12,9.09495e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,1.82254e-12,1.82077e-12,1.81988e-12,1.81943e-12,1.81921e-12,8.18545e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,9.11271e-13,9.10383e-13,9.09939e-13,9.09717e-13,7.7307e-12,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,4.55636e-13,4.55191e-13,4.54969e-13,7.50333e-12,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.27818e-13,2.27596e-13,7.38964e-12,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.13909e-13,7.3328e-12,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,7.30438e-12,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,7.29017e-12,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,7.28306e-12,3.64508e-12,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.27951e-12,3.64153e-12,1.82254e-12,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,7.27773e-12,3.63976e-12,1.82077e-12,9.11271e-13,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,7.27685e-12,3.63887e-12,1.81988e-12,9.10383e-13,4.55636e-13,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,7.2764e-12,3.63842e-12,1.81943e-12,9.09939e-13,4.55191e-13,2.27818e-13,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,7.27618e-12,3.6382e-12,1.81921e-12,9.09717e-13,4.54969e-13,2.27596e-13,1.13909e-13,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,},
-{7.27596e-12,5.45697e-12,4.54747e-12,4.09273e-12,3.86535e-12,3.75167e-12,3.69482e-12,3.6664e-12,3.65219e-12,3.64508e-12,3.64153e-12,3.63976e-12,3.63887e-12,3.63842e-12,3.6382e-12,3.63809e-12,5.45697e-12,3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,1.82254e-12,1.82077e-12,1.81988e-12,1.81943e-12,1.81921e-12,1.8191e-12,4.54747e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,9.11271e-13,9.10383e-13,9.09939e-13,9.09717e-13,9.09606e-13,4.09273e-12,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,4.55636e-13,4.55191e-13,4.54969e-13,4.54858e-13,3.86535e-12,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.27818e-13,2.27596e-13,2.27485e-13,3.75167e-12,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.13909e-13,1.13798e-13,3.69482e-12,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,5.69544e-14,3.6664e-12,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,3.65219e-12,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,3.64508e-12,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,3.64153e-12,1.82254e-12,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.63976e-12,1.82077e-12,9.11271e-13,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,3.63887e-12,1.81988e-12,9.10383e-13,4.55636e-13,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,3.63842e-12,1.81943e-12,9.09939e-13,4.55191e-13,2.27818e-13,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,3.6382e-12,1.81921e-12,9.09717e-13,4.54969e-13,2.27596e-13,1.13909e-13,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,3.63809e-12,1.8191e-12,9.09606e-13,4.54858e-13,2.27485e-13,1.13798e-13,5.69544e-14,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,},
-{3.63798e-12,2.72848e-12,2.27374e-12,2.04636e-12,1.93268e-12,1.87583e-12,1.84741e-12,1.8332e-12,1.82609e-12,1.82254e-12,1.82077e-12,1.81988e-12,1.81943e-12,1.81921e-12,1.8191e-12,1.81904e-12,2.72848e-12,1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,9.11271e-13,9.10383e-13,9.09939e-13,9.09717e-13,9.09606e-13,9.0955e-13,2.27374e-12,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,4.55636e-13,4.55191e-13,4.54969e-13,4.54858e-13,4.54803e-13,2.04636e-12,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.27818e-13,2.27596e-13,2.27485e-13,2.27429e-13,1.93268e-12,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.13909e-13,1.13798e-13,1.13742e-13,1.87583e-12,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,5.69544e-14,5.68989e-14,1.84741e-12,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,2.84772e-14,1.8332e-12,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.82609e-12,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,1.82254e-12,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,1.82077e-12,9.11271e-13,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.81988e-12,9.10383e-13,4.55636e-13,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,1.81943e-12,9.09939e-13,4.55191e-13,2.27818e-13,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,1.81921e-12,9.09717e-13,4.54969e-13,2.27596e-13,1.13909e-13,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,1.8191e-12,9.09606e-13,4.54858e-13,2.27485e-13,1.13798e-13,5.69544e-14,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.81904e-12,9.0955e-13,4.54803e-13,2.27429e-13,1.13742e-13,5.68989e-14,2.84772e-14,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,},
-{1.81899e-12,1.36424e-12,1.13687e-12,1.02318e-12,9.66338e-13,9.37916e-13,9.23706e-13,9.166e-13,9.13047e-13,9.11271e-13,9.10383e-13,9.09939e-13,9.09717e-13,9.09606e-13,9.0955e-13,9.09522e-13,1.36424e-12,9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,4.55636e-13,4.55191e-13,4.54969e-13,4.54858e-13,4.54803e-13,4.54775e-13,1.13687e-12,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.27818e-13,2.27596e-13,2.27485e-13,2.27429e-13,2.27401e-13,1.02318e-12,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.13909e-13,1.13798e-13,1.13742e-13,1.13715e-13,9.66338e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,5.69544e-14,5.68989e-14,5.68712e-14,9.37916e-13,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,2.84772e-14,2.84495e-14,9.23706e-13,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.42386e-14,9.166e-13,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,9.13047e-13,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,9.11271e-13,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,9.10383e-13,4.55636e-13,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.09939e-13,4.55191e-13,2.27818e-13,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,9.09717e-13,4.54969e-13,2.27596e-13,1.13909e-13,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,9.09606e-13,4.54858e-13,2.27485e-13,1.13798e-13,5.69544e-14,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,9.0955e-13,4.54803e-13,2.27429e-13,1.13742e-13,5.68989e-14,2.84772e-14,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,9.09522e-13,4.54775e-13,2.27401e-13,1.13715e-13,5.68712e-14,2.84495e-14,1.42386e-14,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,},
-{9.09495e-13,6.82121e-13,5.68434e-13,5.11591e-13,4.83169e-13,4.68958e-13,4.61853e-13,4.583e-13,4.56524e-13,4.55636e-13,4.55191e-13,4.54969e-13,4.54858e-13,4.54803e-13,4.54775e-13,4.54761e-13,6.82121e-13,4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.27818e-13,2.27596e-13,2.27485e-13,2.27429e-13,2.27401e-13,2.27388e-13,5.68434e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.13909e-13,1.13798e-13,1.13742e-13,1.13715e-13,1.13701e-13,5.11591e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,5.69544e-14,5.68989e-14,5.68712e-14,5.68573e-14,4.83169e-13,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,2.84772e-14,2.84495e-14,2.84356e-14,4.68958e-13,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.42386e-14,1.42247e-14,4.61853e-13,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,7.11931e-15,4.583e-13,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,4.56524e-13,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,4.55636e-13,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,4.55191e-13,2.27818e-13,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.54969e-13,2.27596e-13,1.13909e-13,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,4.54858e-13,2.27485e-13,1.13798e-13,5.69544e-14,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,4.54803e-13,2.27429e-13,1.13742e-13,5.68989e-14,2.84772e-14,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,4.54775e-13,2.27401e-13,1.13715e-13,5.68712e-14,2.84495e-14,1.42386e-14,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,4.54761e-13,2.27388e-13,1.13701e-13,5.68573e-14,2.84356e-14,1.42247e-14,7.11931e-15,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,},
-{4.54747e-13,3.41061e-13,2.84217e-13,2.55795e-13,2.41585e-13,2.34479e-13,2.30926e-13,2.2915e-13,2.28262e-13,2.27818e-13,2.27596e-13,2.27485e-13,2.27429e-13,2.27401e-13,2.27388e-13,2.27381e-13,3.41061e-13,2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.13909e-13,1.13798e-13,1.13742e-13,1.13715e-13,1.13701e-13,1.13694e-13,2.84217e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,5.69544e-14,5.68989e-14,5.68712e-14,5.68573e-14,5.68504e-14,2.55795e-13,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,2.84772e-14,2.84495e-14,2.84356e-14,2.84286e-14,2.41585e-13,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.42386e-14,1.42247e-14,1.42178e-14,2.34479e-13,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,7.11931e-15,7.11237e-15,2.30926e-13,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,3.55965e-15,2.2915e-13,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,2.28262e-13,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,2.27818e-13,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,2.27596e-13,1.13909e-13,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.27485e-13,1.13798e-13,5.69544e-14,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,2.27429e-13,1.13742e-13,5.68989e-14,2.84772e-14,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,2.27401e-13,1.13715e-13,5.68712e-14,2.84495e-14,1.42386e-14,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,2.27388e-13,1.13701e-13,5.68573e-14,2.84356e-14,1.42247e-14,7.11931e-15,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,2.27381e-13,1.13694e-13,5.68504e-14,2.84286e-14,1.42178e-14,7.11237e-15,3.55965e-15,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,},
-{2.27374e-13,1.7053e-13,1.42109e-13,1.27898e-13,1.20792e-13,1.1724e-13,1.15463e-13,1.14575e-13,1.14131e-13,1.13909e-13,1.13798e-13,1.13742e-13,1.13715e-13,1.13701e-13,1.13694e-13,1.1369e-13,1.7053e-13,1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,5.69544e-14,5.68989e-14,5.68712e-14,5.68573e-14,5.68504e-14,5.68469e-14,1.42109e-13,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,2.84772e-14,2.84495e-14,2.84356e-14,2.84286e-14,2.84252e-14,1.27898e-13,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.42386e-14,1.42247e-14,1.42178e-14,1.42143e-14,1.20792e-13,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,7.11931e-15,7.11237e-15,7.1089e-15,1.1724e-13,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,3.55965e-15,3.55618e-15,1.15463e-13,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,1.77983e-15,1.14575e-13,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,1.14131e-13,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,1.13909e-13,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,1.13798e-13,5.69544e-14,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.13742e-13,5.68989e-14,2.84772e-14,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,1.13715e-13,5.68712e-14,2.84495e-14,1.42386e-14,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,1.13701e-13,5.68573e-14,2.84356e-14,1.42247e-14,7.11931e-15,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.13694e-13,5.68504e-14,2.84286e-14,1.42178e-14,7.11237e-15,3.55965e-15,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,1.1369e-13,5.68469e-14,2.84252e-14,1.42143e-14,7.1089e-15,3.55618e-15,1.77983e-15,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,},
-{1.13687e-13,8.52651e-14,7.10543e-14,6.39488e-14,6.03961e-14,5.86198e-14,5.77316e-14,5.72875e-14,5.70655e-14,5.69544e-14,5.68989e-14,5.68712e-14,5.68573e-14,5.68504e-14,5.68469e-14,5.68452e-14,8.52651e-14,5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,2.84772e-14,2.84495e-14,2.84356e-14,2.84286e-14,2.84252e-14,2.84234e-14,7.10543e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.42386e-14,1.42247e-14,1.42178e-14,1.42143e-14,1.42126e-14,6.39488e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,7.11931e-15,7.11237e-15,7.1089e-15,7.10716e-15,6.03961e-14,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,3.55965e-15,3.55618e-15,3.55445e-15,5.86198e-14,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,1.77983e-15,1.77809e-15,5.77316e-14,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,8.89913e-16,5.72875e-14,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,5.70655e-14,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,5.69544e-14,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,5.68989e-14,2.84772e-14,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.68712e-14,2.84495e-14,1.42386e-14,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,5.68573e-14,2.84356e-14,1.42247e-14,7.11931e-15,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,5.68504e-14,2.84286e-14,1.42178e-14,7.11237e-15,3.55965e-15,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,5.68469e-14,2.84252e-14,1.42143e-14,7.1089e-15,3.55618e-15,1.77983e-15,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,5.68452e-14,2.84234e-14,1.42126e-14,7.10716e-15,3.55445e-15,1.77809e-15,8.89913e-16,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,},
-{5.68434e-14,4.26326e-14,3.55271e-14,3.19744e-14,3.01981e-14,2.93099e-14,2.88658e-14,2.86438e-14,2.85327e-14,2.84772e-14,2.84495e-14,2.84356e-14,2.84286e-14,2.84252e-14,2.84234e-14,2.84226e-14,4.26326e-14,2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.42386e-14,1.42247e-14,1.42178e-14,1.42143e-14,1.42126e-14,1.42117e-14,3.55271e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,7.11931e-15,7.11237e-15,7.1089e-15,7.10716e-15,7.10629e-15,3.19744e-14,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,3.55965e-15,3.55618e-15,3.55445e-15,3.55358e-15,3.01981e-14,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,1.77983e-15,1.77809e-15,1.77722e-15,2.93099e-14,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,8.89913e-16,8.89046e-16,2.88658e-14,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,4.44957e-16,2.86438e-14,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.85327e-14,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,2.84772e-14,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,2.84495e-14,1.42386e-14,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.84356e-14,1.42247e-14,7.11931e-15,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,2.84286e-14,1.42178e-14,7.11237e-15,3.55965e-15,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,2.84252e-14,1.42143e-14,7.1089e-15,3.55618e-15,1.77983e-15,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,2.84234e-14,1.42126e-14,7.10716e-15,3.55445e-15,1.77809e-15,8.89913e-16,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.84226e-14,1.42117e-14,7.10629e-15,3.55358e-15,1.77722e-15,8.89046e-16,4.44957e-16,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,},
-{2.84217e-14,2.13163e-14,1.77636e-14,1.59872e-14,1.5099e-14,1.46549e-14,1.44329e-14,1.43219e-14,1.42664e-14,1.42386e-14,1.42247e-14,1.42178e-14,1.42143e-14,1.42126e-14,1.42117e-14,1.42113e-14,2.13163e-14,1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,7.11931e-15,7.11237e-15,7.1089e-15,7.10716e-15,7.10629e-15,7.10586e-15,1.77636e-14,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,3.55965e-15,3.55618e-15,3.55445e-15,3.55358e-15,3.55315e-15,1.59872e-14,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,1.77983e-15,1.77809e-15,1.77722e-15,1.77679e-15,1.5099e-14,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,8.89913e-16,8.89046e-16,8.88612e-16,1.46549e-14,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,4.44957e-16,4.44523e-16,1.44329e-14,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.22478e-16,1.43219e-14,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.42664e-14,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,1.42386e-14,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,1.42247e-14,7.11931e-15,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.42178e-14,7.11237e-15,3.55965e-15,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,1.42143e-14,7.1089e-15,3.55618e-15,1.77983e-15,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,1.42126e-14,7.10716e-15,3.55445e-15,1.77809e-15,8.89913e-16,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.42117e-14,7.10629e-15,3.55358e-15,1.77722e-15,8.89046e-16,4.44957e-16,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.42113e-14,7.10586e-15,3.55315e-15,1.77679e-15,8.88612e-16,4.44523e-16,2.22478e-16,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,},
-{1.42109e-14,1.06581e-14,8.88178e-15,7.99361e-15,7.54952e-15,7.32747e-15,7.21645e-15,7.16094e-15,7.13318e-15,7.11931e-15,7.11237e-15,7.1089e-15,7.10716e-15,7.10629e-15,7.10586e-15,7.10564e-15,1.06581e-14,7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,3.55965e-15,3.55618e-15,3.55445e-15,3.55358e-15,3.55315e-15,3.55293e-15,8.88178e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,1.77983e-15,1.77809e-15,1.77722e-15,1.77679e-15,1.77657e-15,7.99361e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,8.89913e-16,8.89046e-16,8.88612e-16,8.88395e-16,7.54952e-15,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,4.44957e-16,4.44523e-16,4.44306e-16,7.32747e-15,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.22478e-16,2.22261e-16,7.21645e-15,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.11239e-16,7.16094e-15,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,7.13318e-15,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,7.11931e-15,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,7.11237e-15,3.55965e-15,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.1089e-15,3.55618e-15,1.77983e-15,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,7.10716e-15,3.55445e-15,1.77809e-15,8.89913e-16,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,7.10629e-15,3.55358e-15,1.77722e-15,8.89046e-16,4.44957e-16,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,7.10586e-15,3.55315e-15,1.77679e-15,8.88612e-16,4.44523e-16,2.22478e-16,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,7.10564e-15,3.55293e-15,1.77657e-15,8.88395e-16,4.44306e-16,2.22261e-16,1.11239e-16,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,},
-{7.10543e-15,5.32907e-15,4.44089e-15,3.9968e-15,3.77476e-15,3.66374e-15,3.60822e-15,3.58047e-15,3.56659e-15,3.55965e-15,3.55618e-15,3.55445e-15,3.55358e-15,3.55315e-15,3.55293e-15,3.55282e-15,5.32907e-15,3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,1.77983e-15,1.77809e-15,1.77722e-15,1.77679e-15,1.77657e-15,1.77647e-15,4.44089e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,8.89913e-16,8.89046e-16,8.88612e-16,8.88395e-16,8.88287e-16,3.9968e-15,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,4.44957e-16,4.44523e-16,4.44306e-16,4.44198e-16,3.77476e-15,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.22478e-16,2.22261e-16,2.22153e-16,3.66374e-15,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.11239e-16,1.11131e-16,3.60822e-15,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,5.56196e-17,3.58047e-15,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,3.56659e-15,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,3.55965e-15,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,3.55618e-15,1.77983e-15,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.55445e-15,1.77809e-15,8.89913e-16,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,3.55358e-15,1.77722e-15,8.89046e-16,4.44957e-16,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,3.55315e-15,1.77679e-15,8.88612e-16,4.44523e-16,2.22478e-16,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,3.55293e-15,1.77657e-15,8.88395e-16,4.44306e-16,2.22261e-16,1.11239e-16,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,3.55282e-15,1.77647e-15,8.88287e-16,4.44198e-16,2.22153e-16,1.11131e-16,5.56196e-17,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,},
-{3.55271e-15,2.66454e-15,2.22045e-15,1.9984e-15,1.88738e-15,1.83187e-15,1.80411e-15,1.79023e-15,1.7833e-15,1.77983e-15,1.77809e-15,1.77722e-15,1.77679e-15,1.77657e-15,1.77647e-15,1.77641e-15,2.66454e-15,1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,8.89913e-16,8.89046e-16,8.88612e-16,8.88395e-16,8.88287e-16,8.88233e-16,2.22045e-15,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,4.44957e-16,4.44523e-16,4.44306e-16,4.44198e-16,4.44143e-16,1.9984e-15,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.22478e-16,2.22261e-16,2.22153e-16,2.22099e-16,1.88738e-15,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.11239e-16,1.11131e-16,1.11077e-16,1.83187e-15,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,5.56196e-17,5.55654e-17,1.80411e-15,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,2.78098e-17,1.79023e-15,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.7833e-15,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,1.77983e-15,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,1.77809e-15,8.89913e-16,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.77722e-15,8.89046e-16,4.44957e-16,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,1.77679e-15,8.88612e-16,4.44523e-16,2.22478e-16,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,1.77657e-15,8.88395e-16,4.44306e-16,2.22261e-16,1.11239e-16,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,1.77647e-15,8.88287e-16,4.44198e-16,2.22153e-16,1.11131e-16,5.56196e-17,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.77641e-15,8.88233e-16,4.44143e-16,2.22099e-16,1.11077e-16,5.55654e-17,2.78098e-17,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,},
-{1.77636e-15,1.33227e-15,1.11022e-15,9.99201e-16,9.4369e-16,9.15934e-16,9.02056e-16,8.95117e-16,8.91648e-16,8.89913e-16,8.89046e-16,8.88612e-16,8.88395e-16,8.88287e-16,8.88233e-16,8.88206e-16,1.33227e-15,8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,4.44957e-16,4.44523e-16,4.44306e-16,4.44198e-16,4.44143e-16,4.44116e-16,1.11022e-15,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.22478e-16,2.22261e-16,2.22153e-16,2.22099e-16,2.22072e-16,9.99201e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.11239e-16,1.11131e-16,1.11077e-16,1.11049e-16,9.4369e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,5.56196e-17,5.55654e-17,5.55383e-17,9.15934e-16,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,2.78098e-17,2.77827e-17,9.02056e-16,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.39049e-17,8.95117e-16,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,8.91648e-16,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,8.89913e-16,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,8.89046e-16,4.44957e-16,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.88612e-16,4.44523e-16,2.22478e-16,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,8.88395e-16,4.44306e-16,2.22261e-16,1.11239e-16,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,8.88287e-16,4.44198e-16,2.22153e-16,1.11131e-16,5.56196e-17,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,8.88233e-16,4.44143e-16,2.22099e-16,1.11077e-16,5.55654e-17,2.78098e-17,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,8.88206e-16,4.44116e-16,2.22072e-16,1.11049e-16,5.55383e-17,2.77827e-17,1.39049e-17,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,},
-{8.88178e-16,6.66134e-16,5.55112e-16,4.996e-16,4.71845e-16,4.57967e-16,4.51028e-16,4.47559e-16,4.45824e-16,4.44957e-16,4.44523e-16,4.44306e-16,4.44198e-16,4.44143e-16,4.44116e-16,4.44103e-16,6.66134e-16,4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.22478e-16,2.22261e-16,2.22153e-16,2.22099e-16,2.22072e-16,2.22058e-16,5.55112e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.11239e-16,1.11131e-16,1.11077e-16,1.11049e-16,1.11036e-16,4.996e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,5.56196e-17,5.55654e-17,5.55383e-17,5.55247e-17,4.71845e-16,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,2.78098e-17,2.77827e-17,2.77691e-17,4.57967e-16,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.39049e-17,1.38913e-17,4.51028e-16,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,6.95245e-18,4.47559e-16,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,4.45824e-16,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,4.44957e-16,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,4.44523e-16,2.22478e-16,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.44306e-16,2.22261e-16,1.11239e-16,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,4.44198e-16,2.22153e-16,1.11131e-16,5.56196e-17,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,4.44143e-16,2.22099e-16,1.11077e-16,5.55654e-17,2.78098e-17,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,4.44116e-16,2.22072e-16,1.11049e-16,5.55383e-17,2.77827e-17,1.39049e-17,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,4.44103e-16,2.22058e-16,1.11036e-16,5.55247e-17,2.77691e-17,1.38913e-17,6.95245e-18,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,},
-{4.44089e-16,3.33067e-16,2.77556e-16,2.498e-16,2.35922e-16,2.28983e-16,2.25514e-16,2.23779e-16,2.22912e-16,2.22478e-16,2.22261e-16,2.22153e-16,2.22099e-16,2.22072e-16,2.22058e-16,2.22051e-16,3.33067e-16,2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.11239e-16,1.11131e-16,1.11077e-16,1.11049e-16,1.11036e-16,1.11029e-16,2.77556e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,5.56196e-17,5.55654e-17,5.55383e-17,5.55247e-17,5.55179e-17,2.498e-16,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,2.78098e-17,2.77827e-17,2.77691e-17,2.77624e-17,2.35922e-16,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.39049e-17,1.38913e-17,1.38846e-17,2.28983e-16,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,6.95245e-18,6.94567e-18,2.25514e-16,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,3.47622e-18,2.23779e-16,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,2.22912e-16,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,2.22478e-16,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,2.22261e-16,1.11239e-16,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.22153e-16,1.11131e-16,5.56196e-17,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,2.22099e-16,1.11077e-16,5.55654e-17,2.78098e-17,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,2.22072e-16,1.11049e-16,5.55383e-17,2.77827e-17,1.39049e-17,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,2.22058e-16,1.11036e-16,5.55247e-17,2.77691e-17,1.38913e-17,6.95245e-18,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,2.22051e-16,1.11029e-16,5.55179e-17,2.77624e-17,1.38846e-17,6.94567e-18,3.47622e-18,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,},
-{2.22045e-16,1.66533e-16,1.38778e-16,1.249e-16,1.17961e-16,1.14492e-16,1.12757e-16,1.1189e-16,1.11456e-16,1.11239e-16,1.11131e-16,1.11077e-16,1.11049e-16,1.11036e-16,1.11029e-16,1.11026e-16,1.66533e-16,1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,5.56196e-17,5.55654e-17,5.55383e-17,5.55247e-17,5.55179e-17,5.55145e-17,1.38778e-16,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,2.78098e-17,2.77827e-17,2.77691e-17,2.77624e-17,2.7759e-17,1.249e-16,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.39049e-17,1.38913e-17,1.38846e-17,1.38812e-17,1.17961e-16,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,6.95245e-18,6.94567e-18,6.94228e-18,1.14492e-16,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,3.47622e-18,3.47284e-18,1.12757e-16,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,1.73811e-18,1.1189e-16,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,1.11456e-16,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,1.11239e-16,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,1.11131e-16,5.56196e-17,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.11077e-16,5.55654e-17,2.78098e-17,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,1.11049e-16,5.55383e-17,2.77827e-17,1.39049e-17,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,1.11036e-16,5.55247e-17,2.77691e-17,1.38913e-17,6.95245e-18,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.11029e-16,5.55179e-17,2.77624e-17,1.38846e-17,6.94567e-18,3.47622e-18,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,1.11026e-16,5.55145e-17,2.7759e-17,1.38812e-17,6.94228e-18,3.47284e-18,1.73811e-18,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,},
-{1.11022e-16,8.32667e-17,6.93889e-17,6.245e-17,5.89806e-17,5.72459e-17,5.63785e-17,5.59448e-17,5.5728e-17,5.56196e-17,5.55654e-17,5.55383e-17,5.55247e-17,5.55179e-17,5.55145e-17,5.55128e-17,8.32667e-17,5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,2.78098e-17,2.77827e-17,2.77691e-17,2.77624e-17,2.7759e-17,2.77573e-17,6.93889e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.39049e-17,1.38913e-17,1.38846e-17,1.38812e-17,1.38795e-17,6.245e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,6.95245e-18,6.94567e-18,6.94228e-18,6.94059e-18,5.89806e-17,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,3.47622e-18,3.47284e-18,3.47114e-18,5.72459e-17,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,1.73811e-18,1.73642e-18,5.63785e-17,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,8.69056e-19,5.59448e-17,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,5.5728e-17,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,5.56196e-17,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,5.55654e-17,2.78098e-17,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.55383e-17,2.77827e-17,1.39049e-17,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,5.55247e-17,2.77691e-17,1.38913e-17,6.95245e-18,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,5.55179e-17,2.77624e-17,1.38846e-17,6.94567e-18,3.47622e-18,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,5.55145e-17,2.7759e-17,1.38812e-17,6.94228e-18,3.47284e-18,1.73811e-18,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,5.55128e-17,2.77573e-17,1.38795e-17,6.94059e-18,3.47114e-18,1.73642e-18,8.69056e-19,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,},
-{5.55112e-17,4.16334e-17,3.46945e-17,3.1225e-17,2.94903e-17,2.86229e-17,2.81893e-17,2.79724e-17,2.7864e-17,2.78098e-17,2.77827e-17,2.77691e-17,2.77624e-17,2.7759e-17,2.77573e-17,2.77564e-17,4.16334e-17,2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.39049e-17,1.38913e-17,1.38846e-17,1.38812e-17,1.38795e-17,1.38786e-17,3.46945e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,6.95245e-18,6.94567e-18,6.94228e-18,6.94059e-18,6.93974e-18,3.1225e-17,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,3.47622e-18,3.47284e-18,3.47114e-18,3.47029e-18,2.94903e-17,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,1.73811e-18,1.73642e-18,1.73557e-18,2.86229e-17,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,8.69056e-19,8.68209e-19,2.81893e-17,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,4.34528e-19,2.79724e-17,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.7864e-17,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,2.78098e-17,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,2.77827e-17,1.39049e-17,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.77691e-17,1.38913e-17,6.95245e-18,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,2.77624e-17,1.38846e-17,6.94567e-18,3.47622e-18,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,2.7759e-17,1.38812e-17,6.94228e-18,3.47284e-18,1.73811e-18,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,2.77573e-17,1.38795e-17,6.94059e-18,3.47114e-18,1.73642e-18,8.69056e-19,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.77564e-17,1.38786e-17,6.93974e-18,3.47029e-18,1.73557e-18,8.68209e-19,4.34528e-19,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,},
-{2.77556e-17,2.08167e-17,1.73472e-17,1.56125e-17,1.47451e-17,1.43115e-17,1.40946e-17,1.39862e-17,1.3932e-17,1.39049e-17,1.38913e-17,1.38846e-17,1.38812e-17,1.38795e-17,1.38786e-17,1.38782e-17,2.08167e-17,1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,6.95245e-18,6.94567e-18,6.94228e-18,6.94059e-18,6.93974e-18,6.93932e-18,1.73472e-17,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,3.47622e-18,3.47284e-18,3.47114e-18,3.47029e-18,3.46987e-18,1.56125e-17,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,1.73811e-18,1.73642e-18,1.73557e-18,1.73515e-18,1.47451e-17,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,8.69056e-19,8.68209e-19,8.67785e-19,1.43115e-17,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,4.34528e-19,4.34104e-19,1.40946e-17,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.17264e-19,1.39862e-17,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.3932e-17,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,1.39049e-17,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,1.38913e-17,6.95245e-18,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.38846e-17,6.94567e-18,3.47622e-18,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,1.38812e-17,6.94228e-18,3.47284e-18,1.73811e-18,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,1.38795e-17,6.94059e-18,3.47114e-18,1.73642e-18,8.69056e-19,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.38786e-17,6.93974e-18,3.47029e-18,1.73557e-18,8.68209e-19,4.34528e-19,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.38782e-17,6.93932e-18,3.46987e-18,1.73515e-18,8.67785e-19,4.34104e-19,2.17264e-19,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,},
-{1.38778e-17,1.04083e-17,8.67362e-18,7.80626e-18,7.37257e-18,7.15573e-18,7.04731e-18,6.9931e-18,6.966e-18,6.95245e-18,6.94567e-18,6.94228e-18,6.94059e-18,6.93974e-18,6.93932e-18,6.93911e-18,1.04083e-17,6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,3.47622e-18,3.47284e-18,3.47114e-18,3.47029e-18,3.46987e-18,3.46966e-18,8.67362e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,1.73811e-18,1.73642e-18,1.73557e-18,1.73515e-18,1.73494e-18,7.80626e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,8.69056e-19,8.68209e-19,8.67785e-19,8.67573e-19,7.37257e-18,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,4.34528e-19,4.34104e-19,4.33893e-19,7.15573e-18,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.17264e-19,2.17052e-19,7.04731e-18,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.08632e-19,6.9931e-18,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,5.44219e-20,6.966e-18,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,2.73168e-20,6.95245e-18,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.37643e-20,6.94567e-18,3.47622e-18,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,6.98802e-21,6.94228e-18,3.47284e-18,1.73811e-18,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,3.59989e-21,6.94059e-18,3.47114e-18,1.73642e-18,8.69056e-19,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.90582e-21,6.93974e-18,3.47029e-18,1.73557e-18,8.68209e-19,4.34528e-19,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.05879e-21,6.93932e-18,3.46987e-18,1.73515e-18,8.67785e-19,4.34104e-19,2.17264e-19,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,6.35275e-22,6.93911e-18,3.46966e-18,1.73494e-18,8.67573e-19,4.33893e-19,2.17052e-19,1.08632e-19,5.44219e-20,2.73168e-20,1.37643e-20,6.98802e-21,3.59989e-21,1.90582e-21,1.05879e-21,6.35275e-22,4.23516e-22,},
-{6.93889e-18,5.20417e-18,4.33681e-18,3.90313e-18,3.68629e-18,3.57787e-18,3.52366e-18,3.49655e-18,3.483e-18,3.47622e-18,3.47284e-18,3.47114e-18,3.47029e-18,3.46987e-18,3.46966e-18,3.46955e-18,5.20417e-18,3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,1.73811e-18,1.73642e-18,1.73557e-18,1.73515e-18,1.73494e-18,1.73483e-18,4.33681e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,8.69056e-19,8.68209e-19,8.67785e-19,8.67573e-19,8.67468e-19,3.90313e-18,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,4.34528e-19,4.34104e-19,4.33893e-19,4.33787e-19,3.68629e-18,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.17264e-19,2.17052e-19,2.16946e-19,3.57787e-18,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.08632e-19,1.08526e-19,3.52366e-18,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,5.44219e-20,5.4316e-20,3.49655e-18,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,2.73168e-20,2.72109e-20,3.483e-18,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.37643e-20,1.36584e-20,3.47622e-18,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,6.98802e-21,6.88214e-21,3.47284e-18,1.73811e-18,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,3.59989e-21,3.49401e-21,3.47114e-18,1.73642e-18,8.69056e-19,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.90582e-21,1.79995e-21,3.47029e-18,1.73557e-18,8.68209e-19,4.34528e-19,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.05879e-21,9.52912e-22,3.46987e-18,1.73515e-18,8.67785e-19,4.34104e-19,2.17264e-19,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,6.35275e-22,5.29396e-22,3.46966e-18,1.73494e-18,8.67573e-19,4.33893e-19,2.17052e-19,1.08632e-19,5.44219e-20,2.73168e-20,1.37643e-20,6.98802e-21,3.59989e-21,1.90582e-21,1.05879e-21,6.35275e-22,4.23516e-22,3.17637e-22,3.46955e-18,1.73483e-18,8.67468e-19,4.33787e-19,2.16946e-19,1.08526e-19,5.4316e-20,2.72109e-20,1.36584e-20,6.88214e-21,3.49401e-21,1.79995e-21,9.52912e-22,5.29396e-22,3.17637e-22,2.11758e-22,},
-{3.46945e-18,2.60209e-18,2.1684e-18,1.95156e-18,1.84314e-18,1.78893e-18,1.76183e-18,1.74828e-18,1.7415e-18,1.73811e-18,1.73642e-18,1.73557e-18,1.73515e-18,1.73494e-18,1.73483e-18,1.73478e-18,2.60209e-18,1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,8.69056e-19,8.68209e-19,8.67785e-19,8.67573e-19,8.67468e-19,8.67415e-19,2.1684e-18,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,4.34528e-19,4.34104e-19,4.33893e-19,4.33787e-19,4.33734e-19,1.95156e-18,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.17264e-19,2.17052e-19,2.16946e-19,2.16893e-19,1.84314e-18,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.08632e-19,1.08526e-19,1.08473e-19,1.78893e-18,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,5.44219e-20,5.4316e-20,5.4263e-20,1.76183e-18,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,2.73168e-20,2.72109e-20,2.7158e-20,1.74828e-18,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.37643e-20,1.36584e-20,1.36055e-20,1.7415e-18,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,6.98802e-21,6.88214e-21,6.8292e-21,1.73811e-18,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,3.59989e-21,3.49401e-21,3.44107e-21,1.73642e-18,8.69056e-19,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.90582e-21,1.79995e-21,1.74701e-21,1.73557e-18,8.68209e-19,4.34528e-19,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.05879e-21,9.52912e-22,8.99973e-22,1.73515e-18,8.67785e-19,4.34104e-19,2.17264e-19,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,6.35275e-22,5.29396e-22,4.76456e-22,1.73494e-18,8.67573e-19,4.33893e-19,2.17052e-19,1.08632e-19,5.44219e-20,2.73168e-20,1.37643e-20,6.98802e-21,3.59989e-21,1.90582e-21,1.05879e-21,6.35275e-22,4.23516e-22,3.17637e-22,2.64698e-22,1.73483e-18,8.67468e-19,4.33787e-19,2.16946e-19,1.08526e-19,5.4316e-20,2.72109e-20,1.36584e-20,6.88214e-21,3.49401e-21,1.79995e-21,9.52912e-22,5.29396e-22,3.17637e-22,2.11758e-22,1.58819e-22,1.73478e-18,8.67415e-19,4.33734e-19,2.16893e-19,1.08473e-19,5.4263e-20,2.7158e-20,1.36055e-20,6.8292e-21,3.44107e-21,1.74701e-21,8.99973e-22,4.76456e-22,2.64698e-22,1.58819e-22,1.05879e-22,},
-{1.73472e-18,1.30104e-18,1.0842e-18,9.75782e-19,9.21572e-19,8.94467e-19,8.80914e-19,8.74138e-19,8.7075e-19,8.69056e-19,8.68209e-19,8.67785e-19,8.67573e-19,8.67468e-19,8.67415e-19,8.67388e-19,1.30104e-18,8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,4.34528e-19,4.34104e-19,4.33893e-19,4.33787e-19,4.33734e-19,4.33707e-19,1.0842e-18,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.17264e-19,2.17052e-19,2.16946e-19,2.16893e-19,2.16867e-19,9.75782e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.08632e-19,1.08526e-19,1.08473e-19,1.08447e-19,9.21572e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,5.44219e-20,5.4316e-20,5.4263e-20,5.42366e-20,8.94467e-19,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,2.73168e-20,2.72109e-20,2.7158e-20,2.71315e-20,8.80914e-19,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.37643e-20,1.36584e-20,1.36055e-20,1.3579e-20,8.74138e-19,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,6.98802e-21,6.88214e-21,6.8292e-21,6.80273e-21,8.7075e-19,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,3.59989e-21,3.49401e-21,3.44107e-21,3.4146e-21,8.69056e-19,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.90582e-21,1.79995e-21,1.74701e-21,1.72054e-21,8.68209e-19,4.34528e-19,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.05879e-21,9.52912e-22,8.99973e-22,8.73503e-22,8.67785e-19,4.34104e-19,2.17264e-19,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,6.35275e-22,5.29396e-22,4.76456e-22,4.49986e-22,8.67573e-19,4.33893e-19,2.17052e-19,1.08632e-19,5.44219e-20,2.73168e-20,1.37643e-20,6.98802e-21,3.59989e-21,1.90582e-21,1.05879e-21,6.35275e-22,4.23516e-22,3.17637e-22,2.64698e-22,2.38228e-22,8.67468e-19,4.33787e-19,2.16946e-19,1.08526e-19,5.4316e-20,2.72109e-20,1.36584e-20,6.88214e-21,3.49401e-21,1.79995e-21,9.52912e-22,5.29396e-22,3.17637e-22,2.11758e-22,1.58819e-22,1.32349e-22,8.67415e-19,4.33734e-19,2.16893e-19,1.08473e-19,5.4263e-20,2.7158e-20,1.36055e-20,6.8292e-21,3.44107e-21,1.74701e-21,8.99973e-22,4.76456e-22,2.64698e-22,1.58819e-22,1.05879e-22,7.94093e-23,8.67388e-19,4.33707e-19,2.16867e-19,1.08447e-19,5.42366e-20,2.71315e-20,1.3579e-20,6.80273e-21,3.4146e-21,1.72054e-21,8.73503e-22,4.49986e-22,2.38228e-22,1.32349e-22,7.94093e-23,5.29396e-23,},
-{8.67362e-19,6.50521e-19,5.42101e-19,4.87891e-19,4.60786e-19,4.47233e-19,4.40457e-19,4.37069e-19,4.35375e-19,4.34528e-19,4.34104e-19,4.33893e-19,4.33787e-19,4.33734e-19,4.33707e-19,4.33694e-19,6.50521e-19,4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.17264e-19,2.17052e-19,2.16946e-19,2.16893e-19,2.16867e-19,2.16854e-19,5.42101e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.08632e-19,1.08526e-19,1.08473e-19,1.08447e-19,1.08433e-19,4.87891e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,5.44219e-20,5.4316e-20,5.4263e-20,5.42366e-20,5.42233e-20,4.60786e-19,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,2.73168e-20,2.72109e-20,2.7158e-20,2.71315e-20,2.71183e-20,4.47233e-19,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.37643e-20,1.36584e-20,1.36055e-20,1.3579e-20,1.35658e-20,4.40457e-19,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,6.98802e-21,6.88214e-21,6.8292e-21,6.80273e-21,6.7895e-21,4.37069e-19,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,3.59989e-21,3.49401e-21,3.44107e-21,3.4146e-21,3.40137e-21,4.35375e-19,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.90582e-21,1.79995e-21,1.74701e-21,1.72054e-21,1.7073e-21,4.34528e-19,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.05879e-21,9.52912e-22,8.99973e-22,8.73503e-22,8.60268e-22,4.34104e-19,2.17264e-19,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,6.35275e-22,5.29396e-22,4.76456e-22,4.49986e-22,4.36751e-22,4.33893e-19,2.17052e-19,1.08632e-19,5.44219e-20,2.73168e-20,1.37643e-20,6.98802e-21,3.59989e-21,1.90582e-21,1.05879e-21,6.35275e-22,4.23516e-22,3.17637e-22,2.64698e-22,2.38228e-22,2.24993e-22,4.33787e-19,2.16946e-19,1.08526e-19,5.4316e-20,2.72109e-20,1.36584e-20,6.88214e-21,3.49401e-21,1.79995e-21,9.52912e-22,5.29396e-22,3.17637e-22,2.11758e-22,1.58819e-22,1.32349e-22,1.19114e-22,4.33734e-19,2.16893e-19,1.08473e-19,5.4263e-20,2.7158e-20,1.36055e-20,6.8292e-21,3.44107e-21,1.74701e-21,8.99973e-22,4.76456e-22,2.64698e-22,1.58819e-22,1.05879e-22,7.94093e-23,6.61744e-23,4.33707e-19,2.16867e-19,1.08447e-19,5.42366e-20,2.71315e-20,1.3579e-20,6.80273e-21,3.4146e-21,1.72054e-21,8.73503e-22,4.49986e-22,2.38228e-22,1.32349e-22,7.94093e-23,5.29396e-23,3.97047e-23,4.33694e-19,2.16854e-19,1.08433e-19,5.42233e-20,2.71183e-20,1.35658e-20,6.7895e-21,3.40137e-21,1.7073e-21,8.60268e-22,4.36751e-22,2.24993e-22,1.19114e-22,6.61744e-23,3.97047e-23,2.64698e-23,},
-{4.33681e-19,3.25261e-19,2.71051e-19,2.43945e-19,2.30393e-19,2.23617e-19,2.20229e-19,2.18535e-19,2.17687e-19,2.17264e-19,2.17052e-19,2.16946e-19,2.16893e-19,2.16867e-19,2.16854e-19,2.16847e-19,3.25261e-19,2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.08632e-19,1.08526e-19,1.08473e-19,1.08447e-19,1.08433e-19,1.08427e-19,2.71051e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,5.44219e-20,5.4316e-20,5.4263e-20,5.42366e-20,5.42233e-20,5.42167e-20,2.43945e-19,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,2.73168e-20,2.72109e-20,2.7158e-20,2.71315e-20,2.71183e-20,2.71117e-20,2.30393e-19,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.37643e-20,1.36584e-20,1.36055e-20,1.3579e-20,1.35658e-20,1.35591e-20,2.23617e-19,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,6.98802e-21,6.88214e-21,6.8292e-21,6.80273e-21,6.7895e-21,6.78288e-21,2.20229e-19,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,3.59989e-21,3.49401e-21,3.44107e-21,3.4146e-21,3.40137e-21,3.39475e-21,2.18535e-19,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.90582e-21,1.79995e-21,1.74701e-21,1.72054e-21,1.7073e-21,1.70068e-21,2.17687e-19,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.05879e-21,9.52912e-22,8.99973e-22,8.73503e-22,8.60268e-22,8.5365e-22,2.17264e-19,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,6.35275e-22,5.29396e-22,4.76456e-22,4.49986e-22,4.36751e-22,4.30134e-22,2.17052e-19,1.08632e-19,5.44219e-20,2.73168e-20,1.37643e-20,6.98802e-21,3.59989e-21,1.90582e-21,1.05879e-21,6.35275e-22,4.23516e-22,3.17637e-22,2.64698e-22,2.38228e-22,2.24993e-22,2.18376e-22,2.16946e-19,1.08526e-19,5.4316e-20,2.72109e-20,1.36584e-20,6.88214e-21,3.49401e-21,1.79995e-21,9.52912e-22,5.29396e-22,3.17637e-22,2.11758e-22,1.58819e-22,1.32349e-22,1.19114e-22,1.12497e-22,2.16893e-19,1.08473e-19,5.4263e-20,2.7158e-20,1.36055e-20,6.8292e-21,3.44107e-21,1.74701e-21,8.99973e-22,4.76456e-22,2.64698e-22,1.58819e-22,1.05879e-22,7.94093e-23,6.61744e-23,5.9557e-23,2.16867e-19,1.08447e-19,5.42366e-20,2.71315e-20,1.3579e-20,6.80273e-21,3.4146e-21,1.72054e-21,8.73503e-22,4.49986e-22,2.38228e-22,1.32349e-22,7.94093e-23,5.29396e-23,3.97047e-23,3.30872e-23,2.16854e-19,1.08433e-19,5.42233e-20,2.71183e-20,1.35658e-20,6.7895e-21,3.40137e-21,1.7073e-21,8.60268e-22,4.36751e-22,2.24993e-22,1.19114e-22,6.61744e-23,3.97047e-23,2.64698e-23,1.98523e-23,2.16847e-19,1.08427e-19,5.42167e-20,2.71117e-20,1.35591e-20,6.78288e-21,3.39475e-21,1.70068e-21,8.5365e-22,4.30134e-22,2.18376e-22,1.12497e-22,5.9557e-23,3.30872e-23,1.98523e-23,1.32349e-23,},
-{2.1684e-19,1.6263e-19,1.35525e-19,1.21973e-19,1.15196e-19,1.11808e-19,1.10114e-19,1.09267e-19,1.08844e-19,1.08632e-19,1.08526e-19,1.08473e-19,1.08447e-19,1.08433e-19,1.08427e-19,1.08424e-19,1.6263e-19,1.0842e-19,8.13152e-20,6.77626e-20,6.09864e-20,5.75982e-20,5.59042e-20,5.50571e-20,5.46336e-20,5.44219e-20,5.4316e-20,5.4263e-20,5.42366e-20,5.42233e-20,5.42167e-20,5.42134e-20,1.35525e-19,8.13152e-20,5.42101e-20,4.06576e-20,3.38813e-20,3.04932e-20,2.87991e-20,2.79521e-20,2.75286e-20,2.73168e-20,2.72109e-20,2.7158e-20,2.71315e-20,2.71183e-20,2.71117e-20,2.71084e-20,1.21973e-19,6.77626e-20,4.06576e-20,2.71051e-20,2.03288e-20,1.69407e-20,1.52466e-20,1.43996e-20,1.3976e-20,1.37643e-20,1.36584e-20,1.36055e-20,1.3579e-20,1.35658e-20,1.35591e-20,1.35558e-20,1.15196e-19,6.09864e-20,3.38813e-20,2.03288e-20,1.35525e-20,1.01644e-20,8.47033e-21,7.6233e-21,7.19978e-21,6.98802e-21,6.88214e-21,6.8292e-21,6.80273e-21,6.7895e-21,6.78288e-21,6.77957e-21,1.11808e-19,5.75982e-20,3.04932e-20,1.69407e-20,1.01644e-20,6.77626e-21,5.0822e-21,4.23516e-21,3.81165e-21,3.59989e-21,3.49401e-21,3.44107e-21,3.4146e-21,3.40137e-21,3.39475e-21,3.39144e-21,1.10114e-19,5.59042e-20,2.87991e-20,1.52466e-20,8.47033e-21,5.0822e-21,3.38813e-21,2.5411e-21,2.11758e-21,1.90582e-21,1.79995e-21,1.74701e-21,1.72054e-21,1.7073e-21,1.70068e-21,1.69737e-21,1.09267e-19,5.50571e-20,2.79521e-20,1.43996e-20,7.6233e-21,4.23516e-21,2.5411e-21,1.69407e-21,1.27055e-21,1.05879e-21,9.52912e-22,8.99973e-22,8.73503e-22,8.60268e-22,8.5365e-22,8.50342e-22,1.08844e-19,5.46336e-20,2.75286e-20,1.3976e-20,7.19978e-21,3.81165e-21,2.11758e-21,1.27055e-21,8.47033e-22,6.35275e-22,5.29396e-22,4.76456e-22,4.49986e-22,4.36751e-22,4.30134e-22,4.26825e-22,1.08632e-19,5.44219e-20,2.73168e-20,1.37643e-20,6.98802e-21,3.59989e-21,1.90582e-21,1.05879e-21,6.35275e-22,4.23516e-22,3.17637e-22,2.64698e-22,2.38228e-22,2.24993e-22,2.18376e-22,2.15067e-22,1.08526e-19,5.4316e-20,2.72109e-20,1.36584e-20,6.88214e-21,3.49401e-21,1.79995e-21,9.52912e-22,5.29396e-22,3.17637e-22,2.11758e-22,1.58819e-22,1.32349e-22,1.19114e-22,1.12497e-22,1.09188e-22,1.08473e-19,5.4263e-20,2.7158e-20,1.36055e-20,6.8292e-21,3.44107e-21,1.74701e-21,8.99973e-22,4.76456e-22,2.64698e-22,1.58819e-22,1.05879e-22,7.94093e-23,6.61744e-23,5.9557e-23,5.62483e-23,1.08447e-19,5.42366e-20,2.71315e-20,1.3579e-20,6.80273e-21,3.4146e-21,1.72054e-21,8.73503e-22,4.49986e-22,2.38228e-22,1.32349e-22,7.94093e-23,5.29396e-23,3.97047e-23,3.30872e-23,2.97785e-23,1.08433e-19,5.42233e-20,2.71183e-20,1.35658e-20,6.7895e-21,3.40137e-21,1.7073e-21,8.60268e-22,4.36751e-22,2.24993e-22,1.19114e-22,6.61744e-23,3.97047e-23,2.64698e-23,1.98523e-23,1.65436e-23,1.08427e-19,5.42167e-20,2.71117e-20,1.35591e-20,6.78288e-21,3.39475e-21,1.70068e-21,8.5365e-22,4.30134e-22,2.18376e-22,1.12497e-22,5.9557e-23,3.30872e-23,1.98523e-23,1.32349e-23,9.92617e-24,1.08424e-19,5.42134e-20,2.71084e-20,1.35558e-20,6.77957e-21,3.39144e-21,1.69737e-21,8.50342e-22,4.26825e-22,2.15067e-22,1.09188e-22,5.62483e-23,2.97785e-23,1.65436e-23,9.92617e-24,6.61744e-24,},};
