+++ SOURCES/include/hll-criteo/hll.hpp
@@ -1,36 +1,19 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
-#include "hll_raw.hpp"
-#include <bitset>
-
 #ifndef _HLL_H_
 #define _HLL_H_
 
+#include <bitset>
+
+#include "../hll_utils.hpp"
+#include "hll_raw.hpp"
+
 
 struct HLLHdr {
-  char magic[2] = {'H','L'};
-  char format;
+  uint8_t magic[2] = {'H','L'};
+  uint8_t format;
   uint8_t bucketBase;
   uint16_t bucketSparseCount; // Only meaning if format is sparse - only maintained at serialization
-  char padding[2] = {'\0','\0'}; // padding to reach 8 bytes in length
-};
+  uint8_t padding[2] = {'\0','\0'}; // padding to reach 8 bytes in length
+} __packed__;
 
 typedef std::pair<std::unique_ptr<uint8_t[]>, size_t> SizedBuffer;
