+++ SOURCES/include/hll-criteo/hll_raw.hpp
@@ -1,21 +1,5 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
+#ifndef _HLL_RAW_H_
+#define _HLL_RAW_H_
 
 #include <exception>
 #include <algorithm>
@@ -27,13 +11,11 @@ under the License.
 #include <type_traits>
 #include <utility>
 #include <cassert>
+
 #include "bias_corrected_estimate.hpp"
 #include "linear_counting.hpp"
 #include "murmur_hash.hpp"
-
-
-#ifndef _HLL_RAW_H_
-#define _HLL_RAW_H_
+#include "../hll_utils.hpp"
 
 struct SerializationError : public virtual std::runtime_error {
   SerializationError(const char* message) : std::runtime_error(std::string(message)) {}
@@ -298,8 +280,7 @@ public:
     double alpha = getAlpha();
     for (uint64_t i = 0; i < this->getNumberOfBuckets(); i++)
     {
-        uint64_t numerator = 1UL << (synopsis[i]);
-        harmonicMean += 1.0 / numerator;
+        harmonicMean += fast_inv_pow2(synopsis[i]);
     }
     harmonicMean = this->getNumberOfBuckets() / harmonicMean;
     // std::llround returns a long long
@@ -356,8 +337,7 @@ public:
       if (synopsis[i] == 0){
         ++numberOfZeroes;
       }
-      uint64_t numerator = 1UL << (synopsis[i]);
-      harmonicMean += 1.0 / numerator;
+      harmonicMean += fast_inv_pow2(synopsis[i]);
     }
     harmonicMean = this->getNumberOfBuckets() / (harmonicMean + getBeta(numberOfZeroes));
     // std::llround returns a long long
