+++ SOURCES/src/hll-criteo/bias_corrected_estimate.cpp
@@ -1,27 +1,8 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
 #include <algorithm>
 #include <cassert>
 #include <iostream>
 #include <vector>
-#include "bias_corrected_estimate.hpp"
+#include "hll-criteo/bias_corrected_estimate.hpp"
 #include <cmath>
 
 #ifdef HIVE_BUILD
