+++ SOURCES/src/hll-criteo/hll_vertica.cpp
@@ -1,25 +1,6 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
 #include "Vertica.h"
-#include "hll_raw.hpp"
-#include "hll_vertica.hpp"
+#include "hll-criteo/hll_raw.hpp"
+#include "hll-criteo/hll_vertica.hpp"
 
 Format formatCodeToEnum(uint8_t f) {
   Format ret = Format::NORMAL;
