+++ SOURCES/tests/hll-criteo/hll_driver.cpp
@@ -1,28 +1,9 @@
-/*
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements.  See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership.  The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied.  See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
 #include <iostream>
-#include "hll.hpp"
+#include "hll-criteo/hll.hpp"
 
 
 /**
- * A simple driver program to show off how Hll can be used
+ * A simple driver program to show how Hll can be used
  */
 
 int main(int argc, char** argv) {
